Refactor more specs
