+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/contract/precompile/AtomicCryptoTransferHTSSuite.java
@@ -19,7 +19,6 @@ package com.hedera.services.bdd.suites.contract.precompile;
 import static com.hedera.services.bdd.junit.TestTags.SMART_CONTRACT;
 import static com.hedera.services.bdd.spec.HapiPropertySource.asHexedSolidityAddress;
 import static com.hedera.services.bdd.spec.HapiSpec.hapiTest;
-import static com.hedera.services.bdd.spec.HapiSpec.propertyPreservingHapiSpec;
 import static com.hedera.services.bdd.spec.assertions.AccountDetailsAsserts.accountDetailsWith;
 import static com.hedera.services.bdd.spec.assertions.ContractFnResultAsserts.resultWith;
 import static com.hedera.services.bdd.spec.assertions.TransactionRecordAsserts.recordWith;
@@ -55,21 +54,16 @@ import static com.hedera.services.bdd.spec.utilops.UtilVerbs.childRecordsCheck;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.newKeyNamed;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.nftTransfer;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.overriding;
-import static com.hedera.services.bdd.spec.utilops.UtilVerbs.overridingTwo;
+import static com.hedera.services.bdd.spec.utilops.UtilVerbs.sourcing;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.tokenTransferList;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.tokenTransferLists;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.transferList;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.withOpContext;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.wrapIntoTupleArray;
-import static com.hedera.services.bdd.spec.utilops.records.SnapshotMatchMode.ACCEPTED_MONO_GAS_CALCULATION_DIFFERENCE;
-import static com.hedera.services.bdd.spec.utilops.records.SnapshotMatchMode.NONDETERMINISTIC_FUNCTION_PARAMETERS;
-import static com.hedera.services.bdd.spec.utilops.records.SnapshotMatchMode.NONDETERMINISTIC_NONCE;
-import static com.hedera.services.bdd.spec.utilops.records.SnapshotMatchMode.NONDETERMINISTIC_TRANSACTION_FEES;
 import static com.hedera.services.bdd.suites.HapiSuite.DEFAULT_PAYER;
 import static com.hedera.services.bdd.suites.HapiSuite.GENESIS;
 import static com.hedera.services.bdd.suites.HapiSuite.ONE_HBAR;
 import static com.hedera.services.bdd.suites.HapiSuite.ONE_HUNDRED_HBARS;
-import static com.hedera.services.bdd.suites.contract.precompile.ApproveAllowanceSuite.CONTRACTS_PERMITTED_DELEGATE_CALLERS;
 import static com.hedera.services.bdd.suites.utils.MiscEETUtils.metadata;
 import static com.hedera.services.bdd.suites.utils.contracts.precompile.HTSPrecompileResult.htsPrecompileResult;
 import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.AMOUNT_EXCEEDS_ALLOWANCE;
@@ -83,6 +77,7 @@ import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.SUCCESS;
 import com.esaulpaugh.headlong.abi.Tuple;
 import com.hedera.node.app.hapi.utils.ByteStringUtils;
 import com.hedera.services.bdd.junit.HapiTest;
+import com.hedera.services.bdd.junit.LeakyHapiTest;
 import com.hedera.services.bdd.spec.assertions.ContractInfoAsserts;
 import com.hedera.services.bdd.spec.assertions.NonFungibleTransfers;
 import com.hedera.services.bdd.spec.assertions.SomeFungibleTransfers;
@@ -130,184 +125,168 @@ public class AtomicCryptoTransferHTSSuite {
         final var cryptoTransferRevertNoKeyTxn = "cryptoTransferRevertNoKeyTxn";
         final var cryptoTransferRevertBalanceTooLowTxn = "cryptoTransferRevertBalanceTooLowTxn";
 
-        return propertyPreservingHapiSpec("cryptoTransferForHbarOnly")
-                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
-                .given(
-                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
-                        cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(SENDER2).balance(10 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
-                        cryptoCreate(RECEIVER2).balance(2 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
-                        uploadInitCode(CONTRACT),
-                        contractCreate(CONTRACT))
-                .when(
-                        withOpContext((spec, opLog) -> {
-                            final var sender = spec.registry().getAccountID(SENDER);
-                            final var sender2 = spec.registry().getAccountID(SENDER2);
-                            final var receiver = spec.registry().getAccountID(RECEIVER);
-                            final var receiver2 = spec.registry().getAccountID(RECEIVER2);
-                            final var amountToBeSent = 50 * ONE_HBAR;
+        return hapiTest(
+                cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
+                cryptoCreate(SENDER2).balance(10 * ONE_HUNDRED_HBARS),
+                cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
+                cryptoCreate(RECEIVER2).balance(2 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
+                uploadInitCode(CONTRACT),
+                contractCreate(CONTRACT),
+                withOpContext((spec, opLog) -> {
+                    final var sender = spec.registry().getAccountID(SENDER);
+                    final var sender2 = spec.registry().getAccountID(SENDER2);
+                    final var receiver = spec.registry().getAccountID(RECEIVER);
+                    final var receiver2 = spec.registry().getAccountID(RECEIVER2);
+                    final var amountToBeSent = 50 * ONE_HBAR;
 
-                            /*
-                             We will be covering the following test cases
-                             1. Simple hbar transfer between 2 parties
-                             2. When sender does not have the required key
-                             3. When sender's balance is too low
-                             4. Transfer among 3 parties
-                             5. When transfer balances do not add to 0
-                            */
-                            allRunFor(
-                                    spec,
-                                    newKeyNamed(DELEGATE_KEY)
-                                            .shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
-                                    cryptoUpdate(SENDER).key(DELEGATE_KEY),
-                                    cryptoUpdate(RECEIVER).key(DELEGATE_KEY),
-                                    cryptoUpdate(RECEIVER2).key(DELEGATE_KEY),
-                                    // Simple transfer between sender and receiver for 50 *
-                                    // ONE_HBAR
-                                    // should succeed
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(
-                                                                    accountAmount(sender, -amountToBeSent, false),
-                                                                    accountAmount(receiver, amountToBeSent, false))
-                                                            .build(),
-                                                    EMPTY_TUPLE_ARRAY)
-                                            .payingWith(GENESIS)
-                                            .via(cryptoTransferTxn)
-                                            .gas(GAS_TO_OFFER),
-                                    // Simple transfer between sender2 and receiver for 50 *
-                                    // ONE_HBAR
-                                    // should fail because sender2 does not have the right
-                                    // key
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(
-                                                                    accountAmount(sender2, -amountToBeSent, false),
-                                                                    accountAmount(receiver, amountToBeSent, false))
-                                                            .build(),
-                                                    EMPTY_TUPLE_ARRAY)
-                                            .payingWith(GENESIS)
-                                            .via(cryptoTransferRevertNoKeyTxn)
-                                            .gas(GAS_TO_OFFER)
-                                            .hasKnownStatus(CONTRACT_REVERT_EXECUTED),
-                                    // Simple transfer between sender2 and receiver for 1000
-                                    // * ONE_HUNDRED_HBAR
-                                    // should fail because sender does not have enough hbars
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(
-                                                                    accountAmount(
-                                                                            sender, -1000 * ONE_HUNDRED_HBARS, false),
-                                                                    accountAmount(
-                                                                            receiver, 1000 * ONE_HUNDRED_HBARS, false))
-                                                            .build(),
-                                                    EMPTY_TUPLE_ARRAY)
-                                            .payingWith(GENESIS)
-                                            .via(cryptoTransferRevertBalanceTooLowTxn)
-                                            .gas(GAS_TO_OFFER)
-                                            .hasKnownStatus(CONTRACT_REVERT_EXECUTED),
-                                    // Simple transfer between sender, receiver and
-                                    // receiver2 for 50 * ONE_HBAR
-                                    // sender sends 50, receiver get 10 and receiver2 gets
-                                    // 40
-                                    // should succeed
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(
-                                                                    accountAmount(sender, -amountToBeSent, false),
-                                                                    accountAmount(
-                                                                            receiver,
-                                                                            amountToBeSent - (10 * ONE_HBAR),
-                                                                            false),
-                                                                    accountAmount(
-                                                                            receiver2,
-                                                                            amountToBeSent - (40 * ONE_HBAR),
-                                                                            false))
-                                                            .build(),
-                                                    EMPTY_TUPLE_ARRAY)
-                                            .payingWith(GENESIS)
-                                            .via(cryptoTransferMultiTxn)
-                                            .gas(GAS_TO_OFFER),
-                                    // Simple transfer between sender, receiver and
-                                    // receiver2 for 50 * ONE_HBAR
-                                    // sender sends 50, receiver get 5 and receiver2 gets 40
-                                    // should fail because total does not add to 0
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(
-                                                                    accountAmount(sender, -amountToBeSent, false),
-                                                                    accountAmount(
-                                                                            receiver,
-                                                                            amountToBeSent - (5 * ONE_HBAR),
-                                                                            false),
-                                                                    accountAmount(
-                                                                            receiver2,
-                                                                            amountToBeSent - (40 * ONE_HBAR),
-                                                                            false))
-                                                            .build(),
-                                                    EMPTY_TUPLE_ARRAY)
-                                            .payingWith(GENESIS)
-                                            .via(cryptoTransferRevertTxn)
-                                            .gas(GAS_TO_OFFER)
-                                            .hasKnownStatus(CONTRACT_REVERT_EXECUTED));
-                        }),
-                        getTxnRecord(cryptoTransferTxn).andAllChildRecords().logged())
-                .then(
-                        getAccountBalance(SENDER).hasTinyBars(900 * ONE_HBAR),
-                        getAccountBalance(RECEIVER).hasTinyBars(290 * ONE_HBAR),
-                        getAccountBalance(RECEIVER2).hasTinyBars(210 * ONE_HBAR),
-                        childRecordsCheck(
-                                cryptoTransferTxn,
-                                SUCCESS,
-                                recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(SUCCESS)))
-                                        .transfers(including(tinyBarsFromTo(SENDER, RECEIVER, 50 * ONE_HBAR)))),
-                        childRecordsCheck(
-                                cryptoTransferRevertNoKeyTxn,
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith()
-                                        .status(SPENDER_DOES_NOT_HAVE_ALLOWANCE)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(htsPrecompileResult()
-                                                        .withStatus(SPENDER_DOES_NOT_HAVE_ALLOWANCE)))),
-                        childRecordsCheck(
-                                cryptoTransferRevertBalanceTooLowTxn,
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith()
-                                        .status(INSUFFICIENT_ACCOUNT_BALANCE)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(htsPrecompileResult()
-                                                        .withStatus(INSUFFICIENT_ACCOUNT_BALANCE)))),
-                        childRecordsCheck(
-                                cryptoTransferMultiTxn,
-                                SUCCESS,
-                                recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(SUCCESS)))),
-                        childRecordsCheck(
-                                cryptoTransferRevertTxn,
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith()
-                                        .status(INVALID_ACCOUNT_AMOUNTS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(INVALID_ACCOUNT_AMOUNTS)))));
+                    /*
+                     We will be covering the following test cases
+                     1. Simple hbar transfer between 2 parties
+                     2. When sender does not have the required key
+                     3. When sender's balance is too low
+                     4. Transfer among 3 parties
+                     5. When transfer balances do not add to 0
+                    */
+                    allRunFor(
+                            spec,
+                            newKeyNamed(DELEGATE_KEY).shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
+                            cryptoUpdate(SENDER).key(DELEGATE_KEY),
+                            cryptoUpdate(RECEIVER).key(DELEGATE_KEY),
+                            cryptoUpdate(RECEIVER2).key(DELEGATE_KEY),
+                            // Simple transfer between sender and receiver for 50 *
+                            // ONE_HBAR
+                            // should succeed
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(
+                                                            accountAmount(sender, -amountToBeSent, false),
+                                                            accountAmount(receiver, amountToBeSent, false))
+                                                    .build(),
+                                            EMPTY_TUPLE_ARRAY)
+                                    .payingWith(GENESIS)
+                                    .via(cryptoTransferTxn)
+                                    .gas(GAS_TO_OFFER),
+                            // Simple transfer between sender2 and receiver for 50 *
+                            // ONE_HBAR
+                            // should fail because sender2 does not have the right
+                            // key
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(
+                                                            accountAmount(sender2, -amountToBeSent, false),
+                                                            accountAmount(receiver, amountToBeSent, false))
+                                                    .build(),
+                                            EMPTY_TUPLE_ARRAY)
+                                    .payingWith(GENESIS)
+                                    .via(cryptoTransferRevertNoKeyTxn)
+                                    .gas(GAS_TO_OFFER)
+                                    .hasKnownStatus(CONTRACT_REVERT_EXECUTED),
+                            // Simple transfer between sender2 and receiver for 1000
+                            // * ONE_HUNDRED_HBAR
+                            // should fail because sender does not have enough hbars
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(
+                                                            accountAmount(sender, -1000 * ONE_HUNDRED_HBARS, false),
+                                                            accountAmount(receiver, 1000 * ONE_HUNDRED_HBARS, false))
+                                                    .build(),
+                                            EMPTY_TUPLE_ARRAY)
+                                    .payingWith(GENESIS)
+                                    .via(cryptoTransferRevertBalanceTooLowTxn)
+                                    .gas(GAS_TO_OFFER)
+                                    .hasKnownStatus(CONTRACT_REVERT_EXECUTED),
+                            // Simple transfer between sender, receiver and
+                            // receiver2 for 50 * ONE_HBAR
+                            // sender sends 50, receiver get 10 and receiver2 gets
+                            // 40
+                            // should succeed
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(
+                                                            accountAmount(sender, -amountToBeSent, false),
+                                                            accountAmount(
+                                                                    receiver, amountToBeSent - (10 * ONE_HBAR), false),
+                                                            accountAmount(
+                                                                    receiver2, amountToBeSent - (40 * ONE_HBAR), false))
+                                                    .build(),
+                                            EMPTY_TUPLE_ARRAY)
+                                    .payingWith(GENESIS)
+                                    .via(cryptoTransferMultiTxn)
+                                    .gas(GAS_TO_OFFER),
+                            // Simple transfer between sender, receiver and
+                            // receiver2 for 50 * ONE_HBAR
+                            // sender sends 50, receiver get 5 and receiver2 gets 40
+                            // should fail because total does not add to 0
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(
+                                                            accountAmount(sender, -amountToBeSent, false),
+                                                            accountAmount(
+                                                                    receiver, amountToBeSent - (5 * ONE_HBAR), false),
+                                                            accountAmount(
+                                                                    receiver2, amountToBeSent - (40 * ONE_HBAR), false))
+                                                    .build(),
+                                            EMPTY_TUPLE_ARRAY)
+                                    .payingWith(GENESIS)
+                                    .via(cryptoTransferRevertTxn)
+                                    .gas(GAS_TO_OFFER)
+                                    .hasKnownStatus(CONTRACT_REVERT_EXECUTED));
+                }),
+                getTxnRecord(cryptoTransferTxn).andAllChildRecords().logged(),
+                getAccountBalance(SENDER).hasTinyBars(900 * ONE_HBAR),
+                getAccountBalance(RECEIVER).hasTinyBars(290 * ONE_HBAR),
+                getAccountBalance(RECEIVER2).hasTinyBars(210 * ONE_HBAR),
+                childRecordsCheck(
+                        cryptoTransferTxn,
+                        SUCCESS,
+                        recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SUCCESS)))
+                                .transfers(including(tinyBarsFromTo(SENDER, RECEIVER, 50 * ONE_HBAR)))),
+                childRecordsCheck(
+                        cryptoTransferRevertNoKeyTxn,
+                        CONTRACT_REVERT_EXECUTED,
+                        recordWith()
+                                .status(SPENDER_DOES_NOT_HAVE_ALLOWANCE)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SPENDER_DOES_NOT_HAVE_ALLOWANCE)))),
+                childRecordsCheck(
+                        cryptoTransferRevertBalanceTooLowTxn,
+                        CONTRACT_REVERT_EXECUTED,
+                        recordWith()
+                                .status(INSUFFICIENT_ACCOUNT_BALANCE)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(INSUFFICIENT_ACCOUNT_BALANCE)))),
+                childRecordsCheck(
+                        cryptoTransferMultiTxn,
+                        SUCCESS,
+                        recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SUCCESS)))),
+                childRecordsCheck(
+                        cryptoTransferRevertTxn,
+                        CONTRACT_REVERT_EXECUTED,
+                        recordWith()
+                                .status(INVALID_ACCOUNT_AMOUNTS)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(INVALID_ACCOUNT_AMOUNTS)))));
     }
 
     @HapiTest
@@ -315,105 +294,93 @@ public class AtomicCryptoTransferHTSSuite {
         final var cryptoTransferTxnForFungible = "cryptoTransferTxnForFungible";
         final var cryptoTransferRevertNoKeyTxn = "cryptoTransferRevertNoKeyTxn";
 
-        return propertyPreservingHapiSpec(
-                        "cryptoTransferForFungibleTokenOnly",
-                        NONDETERMINISTIC_FUNCTION_PARAMETERS,
-                        NONDETERMINISTIC_TRANSACTION_FEES,
-                        NONDETERMINISTIC_NONCE)
-                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
-                .given(
-                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
-                        cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(SENDER2).balance(10 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
-                        cryptoCreate(TOKEN_TREASURY),
-                        tokenCreate(FUNGIBLE_TOKEN)
-                                .tokenType(TokenType.FUNGIBLE_COMMON)
-                                .initialSupply(TOTAL_SUPPLY)
-                                .treasury(TOKEN_TREASURY),
-                        tokenAssociate(SENDER, List.of(FUNGIBLE_TOKEN)),
-                        tokenAssociate(SENDER2, List.of(FUNGIBLE_TOKEN)),
-                        tokenAssociate(RECEIVER, List.of(FUNGIBLE_TOKEN)),
-                        cryptoTransfer(moving(200, FUNGIBLE_TOKEN).between(TOKEN_TREASURY, SENDER)),
-                        uploadInitCode(CONTRACT),
-                        contractCreate(CONTRACT).maxAutomaticTokenAssociations(1),
-                        getContractInfo(CONTRACT)
-                                .has(ContractInfoAsserts.contractWith().maxAutoAssociations(1))
-                                .logged())
-                .when(
-                        withOpContext((spec, opLog) -> {
-                            final var token = spec.registry().getTokenID(FUNGIBLE_TOKEN);
-                            final var sender = spec.registry().getAccountID(SENDER);
-                            final var sender2 = spec.registry().getAccountID(SENDER2);
-                            final var receiver = spec.registry().getAccountID(RECEIVER);
-                            final var amountToBeSent = 50L;
+        return hapiTest(
+                cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
+                cryptoCreate(SENDER2).balance(10 * ONE_HUNDRED_HBARS),
+                cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
+                cryptoCreate(TOKEN_TREASURY),
+                tokenCreate(FUNGIBLE_TOKEN)
+                        .tokenType(TokenType.FUNGIBLE_COMMON)
+                        .initialSupply(TOTAL_SUPPLY)
+                        .treasury(TOKEN_TREASURY),
+                tokenAssociate(SENDER, List.of(FUNGIBLE_TOKEN)),
+                tokenAssociate(SENDER2, List.of(FUNGIBLE_TOKEN)),
+                tokenAssociate(RECEIVER, List.of(FUNGIBLE_TOKEN)),
+                cryptoTransfer(moving(200, FUNGIBLE_TOKEN).between(TOKEN_TREASURY, SENDER)),
+                uploadInitCode(CONTRACT),
+                contractCreate(CONTRACT).maxAutomaticTokenAssociations(1),
+                getContractInfo(CONTRACT)
+                        .has(ContractInfoAsserts.contractWith().maxAutoAssociations(1))
+                        .logged(),
+                withOpContext((spec, opLog) -> {
+                    final var token = spec.registry().getTokenID(FUNGIBLE_TOKEN);
+                    final var sender = spec.registry().getAccountID(SENDER);
+                    final var sender2 = spec.registry().getAccountID(SENDER2);
+                    final var receiver = spec.registry().getAccountID(RECEIVER);
+                    final var amountToBeSent = 50L;
 
-                            allRunFor(
-                                    spec,
-                                    newKeyNamed(DELEGATE_KEY)
-                                            .shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
-                                    cryptoUpdate(SENDER).key(DELEGATE_KEY),
-                                    cryptoUpdate(RECEIVER).key(DELEGATE_KEY),
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(EMPTY_TUPLE_ARRAY)
-                                                            .build(),
-                                                    wrapIntoTupleArray(tokenTransferList()
-                                                            .forToken(token)
-                                                            .withAccountAmounts(
-                                                                    accountAmount(sender, -amountToBeSent, false),
-                                                                    accountAmount(receiver, amountToBeSent, false))
-                                                            .build()))
-                                            .payingWith(GENESIS)
-                                            .via(cryptoTransferTxnForFungible)
-                                            .gas(GAS_TO_OFFER),
-                                    // Ensure that the transfer fails when the sender does not have the correct key
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(EMPTY_TUPLE_ARRAY)
-                                                            .build(),
-                                                    wrapIntoTupleArray(tokenTransferList()
-                                                            .forToken(token)
-                                                            .withAccountAmounts(
-                                                                    accountAmount(sender2, -amountToBeSent, false),
-                                                                    accountAmount(receiver, amountToBeSent, false))
-                                                            .build()))
-                                            .payingWith(GENESIS)
-                                            .via(cryptoTransferRevertNoKeyTxn)
-                                            .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
-                                            .gas(GAS_TO_OFFER));
-                        }),
-                        getTxnRecord(cryptoTransferTxnForFungible)
-                                .andAllChildRecords()
-                                .logged())
-                .then(
-                        getTokenInfo(FUNGIBLE_TOKEN).hasTotalSupply(TOTAL_SUPPLY),
-                        getAccountBalance(RECEIVER).hasTokenBalance(FUNGIBLE_TOKEN, 50),
-                        getAccountBalance(SENDER).hasTokenBalance(FUNGIBLE_TOKEN, 150),
-                        getTokenInfo(FUNGIBLE_TOKEN).logged(),
-                        childRecordsCheck(
-                                cryptoTransferTxnForFungible,
-                                SUCCESS,
-                                recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(SUCCESS)))
-                                        .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
-                                                .including(FUNGIBLE_TOKEN, SENDER, -50)
-                                                .including(FUNGIBLE_TOKEN, RECEIVER, 50))),
-                        childRecordsCheck(
-                                cryptoTransferRevertNoKeyTxn,
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith()
-                                        .status(SPENDER_DOES_NOT_HAVE_ALLOWANCE)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(htsPrecompileResult()
-                                                        .withStatus(SPENDER_DOES_NOT_HAVE_ALLOWANCE)))));
+                    allRunFor(
+                            spec,
+                            newKeyNamed(DELEGATE_KEY).shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
+                            cryptoUpdate(SENDER).key(DELEGATE_KEY),
+                            cryptoUpdate(RECEIVER).key(DELEGATE_KEY),
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(EMPTY_TUPLE_ARRAY)
+                                                    .build(),
+                                            wrapIntoTupleArray(tokenTransferList()
+                                                    .forToken(token)
+                                                    .withAccountAmounts(
+                                                            accountAmount(sender, -amountToBeSent, false),
+                                                            accountAmount(receiver, amountToBeSent, false))
+                                                    .build()))
+                                    .payingWith(GENESIS)
+                                    .via(cryptoTransferTxnForFungible)
+                                    .gas(GAS_TO_OFFER),
+                            // Ensure that the transfer fails when the sender does not have the correct key
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(EMPTY_TUPLE_ARRAY)
+                                                    .build(),
+                                            wrapIntoTupleArray(tokenTransferList()
+                                                    .forToken(token)
+                                                    .withAccountAmounts(
+                                                            accountAmount(sender2, -amountToBeSent, false),
+                                                            accountAmount(receiver, amountToBeSent, false))
+                                                    .build()))
+                                    .payingWith(GENESIS)
+                                    .via(cryptoTransferRevertNoKeyTxn)
+                                    .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
+                                    .gas(GAS_TO_OFFER));
+                }),
+                getTxnRecord(cryptoTransferTxnForFungible).andAllChildRecords(),
+                getTokenInfo(FUNGIBLE_TOKEN).hasTotalSupply(TOTAL_SUPPLY),
+                getAccountBalance(RECEIVER).hasTokenBalance(FUNGIBLE_TOKEN, 50),
+                getAccountBalance(SENDER).hasTokenBalance(FUNGIBLE_TOKEN, 150),
+                getTokenInfo(FUNGIBLE_TOKEN).logged(),
+                childRecordsCheck(
+                        cryptoTransferTxnForFungible,
+                        SUCCESS,
+                        recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SUCCESS)))
+                                .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
+                                        .including(FUNGIBLE_TOKEN, SENDER, -50)
+                                        .including(FUNGIBLE_TOKEN, RECEIVER, 50))),
+                childRecordsCheck(
+                        cryptoTransferRevertNoKeyTxn,
+                        CONTRACT_REVERT_EXECUTED,
+                        recordWith()
+                                .status(SPENDER_DOES_NOT_HAVE_ALLOWANCE)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SPENDER_DOES_NOT_HAVE_ALLOWANCE)))));
     }
 
     @HapiTest
@@ -421,89 +388,75 @@ public class AtomicCryptoTransferHTSSuite {
         final var cryptoTransferTxnForFungible = "cryptoTransferTxnForFungible";
         final var FEE_TOKEN = "FeeToken";
 
-        return propertyPreservingHapiSpec(
-                        "cryptoTransferForFungibleTokenWithFees",
-                        NONDETERMINISTIC_FUNCTION_PARAMETERS,
-                        NONDETERMINISTIC_TRANSACTION_FEES,
-                        NONDETERMINISTIC_NONCE)
-                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
-                .given(
-                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
-                        cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
-                        cryptoCreate(TOKEN_TREASURY),
-                        tokenCreate(FEE_TOKEN)
-                                .tokenType(TokenType.FUNGIBLE_COMMON)
-                                .initialSupply(TOTAL_SUPPLY)
-                                .treasury(TOKEN_TREASURY),
-                        tokenCreate(FUNGIBLE_TOKEN)
-                                .tokenType(TokenType.FUNGIBLE_COMMON)
-                                .initialSupply(TOTAL_SUPPLY)
-                                .treasury(TOKEN_TREASURY)
-                                .withCustom(fractionalFee(1, 10L, 1L, OptionalLong.empty(), TOKEN_TREASURY))
-                                .withCustom(fixedHtsFee(1L, FEE_TOKEN, TOKEN_TREASURY)),
-                        tokenAssociate(SENDER, List.of(FUNGIBLE_TOKEN, FEE_TOKEN)),
-                        tokenAssociate(RECEIVER, List.of(FUNGIBLE_TOKEN)),
-                        cryptoTransfer(moving(10, FEE_TOKEN).between(TOKEN_TREASURY, SENDER)),
-                        cryptoTransfer(moving(200, FUNGIBLE_TOKEN).between(TOKEN_TREASURY, SENDER)),
-                        uploadInitCode(CONTRACT),
-                        contractCreate(CONTRACT).maxAutomaticTokenAssociations(1),
-                        getContractInfo(CONTRACT)
-                                .has(ContractInfoAsserts.contractWith().maxAutoAssociations(1))
-                                .logged())
-                .when(
-                        withOpContext((spec, opLog) -> {
-                            final var token = spec.registry().getTokenID(FUNGIBLE_TOKEN);
-                            final var sender = spec.registry().getAccountID(SENDER);
-                            final var receiver = spec.registry().getAccountID(RECEIVER);
-                            final var amountToBeSent = 50L;
+        return hapiTest(
+                cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
+                cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
+                cryptoCreate(TOKEN_TREASURY),
+                tokenCreate(FEE_TOKEN)
+                        .tokenType(TokenType.FUNGIBLE_COMMON)
+                        .initialSupply(TOTAL_SUPPLY)
+                        .treasury(TOKEN_TREASURY),
+                tokenCreate(FUNGIBLE_TOKEN)
+                        .tokenType(TokenType.FUNGIBLE_COMMON)
+                        .initialSupply(TOTAL_SUPPLY)
+                        .treasury(TOKEN_TREASURY)
+                        .withCustom(fractionalFee(1, 10L, 1L, OptionalLong.empty(), TOKEN_TREASURY))
+                        .withCustom(fixedHtsFee(1L, FEE_TOKEN, TOKEN_TREASURY)),
+                tokenAssociate(SENDER, List.of(FUNGIBLE_TOKEN, FEE_TOKEN)),
+                tokenAssociate(RECEIVER, List.of(FUNGIBLE_TOKEN)),
+                cryptoTransfer(moving(10, FEE_TOKEN).between(TOKEN_TREASURY, SENDER)),
+                cryptoTransfer(moving(200, FUNGIBLE_TOKEN).between(TOKEN_TREASURY, SENDER)),
+                uploadInitCode(CONTRACT),
+                contractCreate(CONTRACT).maxAutomaticTokenAssociations(1),
+                getContractInfo(CONTRACT).has(ContractInfoAsserts.contractWith().maxAutoAssociations(1)),
+                withOpContext((spec, opLog) -> {
+                    final var token = spec.registry().getTokenID(FUNGIBLE_TOKEN);
+                    final var sender = spec.registry().getAccountID(SENDER);
+                    final var receiver = spec.registry().getAccountID(RECEIVER);
+                    final var amountToBeSent = 50L;
 
-                            allRunFor(
-                                    spec,
-                                    newKeyNamed(DELEGATE_KEY)
-                                            .shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
-                                    cryptoUpdate(SENDER).key(DELEGATE_KEY),
-                                    cryptoUpdate(RECEIVER).key(DELEGATE_KEY),
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(EMPTY_TUPLE_ARRAY)
-                                                            .build(),
-                                                    wrapIntoTupleArray(tokenTransferList()
-                                                            .forToken(token)
-                                                            .withAccountAmounts(
-                                                                    accountAmount(sender, -amountToBeSent, false),
-                                                                    accountAmount(receiver, amountToBeSent, false))
-                                                            .build()))
-                                            .payingWith(GENESIS)
-                                            .via(cryptoTransferTxnForFungible)
-                                            .gas(GAS_TO_OFFER));
-                        }),
-                        getTxnRecord(cryptoTransferTxnForFungible)
-                                .andAllChildRecords()
-                                .logged())
-                .then(
-                        getTokenInfo(FUNGIBLE_TOKEN).hasTotalSupply(TOTAL_SUPPLY),
-                        getAccountBalance(RECEIVER).hasTokenBalance(FUNGIBLE_TOKEN, 45),
-                        getAccountBalance(SENDER).hasTokenBalance(FUNGIBLE_TOKEN, 150),
-                        getTokenInfo(FUNGIBLE_TOKEN).logged(),
-                        childRecordsCheck(
-                                cryptoTransferTxnForFungible,
-                                SUCCESS,
-                                recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(SUCCESS)))
-                                        .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
-                                                .including(FUNGIBLE_TOKEN, SENDER, -50)
-                                                .including(FUNGIBLE_TOKEN, RECEIVER, 45)
-                                                .including(FUNGIBLE_TOKEN, TOKEN_TREASURY, 5))
-                                        .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
-                                                .including(FEE_TOKEN, SENDER, -1)
-                                                .including(FEE_TOKEN, TOKEN_TREASURY, 1))
-                                        .assessedCustomFeeCount(2)));
+                    allRunFor(
+                            spec,
+                            newKeyNamed(DELEGATE_KEY).shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
+                            cryptoUpdate(SENDER).key(DELEGATE_KEY),
+                            cryptoUpdate(RECEIVER).key(DELEGATE_KEY),
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(EMPTY_TUPLE_ARRAY)
+                                                    .build(),
+                                            wrapIntoTupleArray(tokenTransferList()
+                                                    .forToken(token)
+                                                    .withAccountAmounts(
+                                                            accountAmount(sender, -amountToBeSent, false),
+                                                            accountAmount(receiver, amountToBeSent, false))
+                                                    .build()))
+                                    .payingWith(GENESIS)
+                                    .via(cryptoTransferTxnForFungible)
+                                    .gas(GAS_TO_OFFER));
+                }),
+                getTxnRecord(cryptoTransferTxnForFungible).andAllChildRecords(),
+                getTokenInfo(FUNGIBLE_TOKEN).hasTotalSupply(TOTAL_SUPPLY),
+                getAccountBalance(RECEIVER).hasTokenBalance(FUNGIBLE_TOKEN, 45),
+                getAccountBalance(SENDER).hasTokenBalance(FUNGIBLE_TOKEN, 150),
+                getTokenInfo(FUNGIBLE_TOKEN).logged(),
+                childRecordsCheck(
+                        cryptoTransferTxnForFungible,
+                        SUCCESS,
+                        recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SUCCESS)))
+                                .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
+                                        .including(FUNGIBLE_TOKEN, SENDER, -50)
+                                        .including(FUNGIBLE_TOKEN, RECEIVER, 45)
+                                        .including(FUNGIBLE_TOKEN, TOKEN_TREASURY, 5))
+                                .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
+                                        .including(FEE_TOKEN, SENDER, -1)
+                                        .including(FEE_TOKEN, TOKEN_TREASURY, 1))
+                                .assessedCustomFeeCount(2)));
     }
 
     @HapiTest
@@ -511,93 +464,83 @@ public class AtomicCryptoTransferHTSSuite {
         final var cryptoTransferTxnForNonFungible = "cryptoTransferTxnForNonFungible";
         final var FEE_TOKEN = "FeeToken";
 
-        return propertyPreservingHapiSpec(
-                        "cryptoTransferForNFTWithFees",
-                        NONDETERMINISTIC_FUNCTION_PARAMETERS,
-                        NONDETERMINISTIC_TRANSACTION_FEES,
-                        NONDETERMINISTIC_NONCE)
-                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
-                .given(
-                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
-                        newKeyNamed(MULTI_KEY),
-                        cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(TOKEN_TREASURY),
-                        tokenCreate(FEE_TOKEN)
-                                .tokenType(TokenType.FUNGIBLE_COMMON)
-                                .initialSupply(TOTAL_SUPPLY)
-                                .treasury(TOKEN_TREASURY),
-                        tokenCreate(NFT_TOKEN)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .adminKey(MULTI_KEY)
-                                .supplyKey(MULTI_KEY)
-                                .supplyType(TokenSupplyType.INFINITE)
-                                .initialSupply(0)
-                                .treasury(TOKEN_TREASURY)
-                                .withCustom(royaltyFeeWithFallback(
-                                        1L, 10L, fixedHtsFeeInheritingRoyaltyCollector(1, FEE_TOKEN), TOKEN_TREASURY)),
-                        mintToken(NFT_TOKEN, List.of(metadata("firstMemo"), metadata("secondMemo"))),
-                        tokenAssociate(SENDER, List.of(NFT_TOKEN, FEE_TOKEN)),
-                        tokenAssociate(RECEIVER, List.of(NFT_TOKEN, FEE_TOKEN)),
-                        cryptoTransfer(moving(10, FEE_TOKEN).between(TOKEN_TREASURY, RECEIVER)),
-                        cryptoTransfer(TokenMovement.movingUnique(NFT_TOKEN, 1).between(TOKEN_TREASURY, SENDER))
-                                .payingWith(SENDER),
-                        uploadInitCode(CONTRACT),
-                        contractCreate(CONTRACT).maxAutomaticTokenAssociations(1),
-                        getContractInfo(CONTRACT)
-                                .has(ContractInfoAsserts.contractWith().maxAutoAssociations(1))
-                                .logged())
-                .when(
-                        withOpContext((spec, opLog) -> {
-                            final var token = spec.registry().getTokenID(NFT_TOKEN);
-                            final var sender = spec.registry().getAccountID(SENDER);
-                            final var receiver = spec.registry().getAccountID(RECEIVER);
-                            final var amountToBeSent = 50L;
+        return hapiTest(
+                newKeyNamed(MULTI_KEY),
+                cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
+                cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS),
+                cryptoCreate(TOKEN_TREASURY),
+                tokenCreate(FEE_TOKEN)
+                        .tokenType(TokenType.FUNGIBLE_COMMON)
+                        .initialSupply(TOTAL_SUPPLY)
+                        .treasury(TOKEN_TREASURY),
+                tokenCreate(NFT_TOKEN)
+                        .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
+                        .adminKey(MULTI_KEY)
+                        .supplyKey(MULTI_KEY)
+                        .supplyType(TokenSupplyType.INFINITE)
+                        .initialSupply(0)
+                        .treasury(TOKEN_TREASURY)
+                        .withCustom(royaltyFeeWithFallback(
+                                1L, 10L, fixedHtsFeeInheritingRoyaltyCollector(1, FEE_TOKEN), TOKEN_TREASURY)),
+                mintToken(NFT_TOKEN, List.of(metadata("firstMemo"), metadata("secondMemo"))),
+                tokenAssociate(SENDER, List.of(NFT_TOKEN, FEE_TOKEN)),
+                tokenAssociate(RECEIVER, List.of(NFT_TOKEN, FEE_TOKEN)),
+                cryptoTransfer(moving(10, FEE_TOKEN).between(TOKEN_TREASURY, RECEIVER)),
+                cryptoTransfer(TokenMovement.movingUnique(NFT_TOKEN, 1).between(TOKEN_TREASURY, SENDER))
+                        .payingWith(SENDER),
+                uploadInitCode(CONTRACT),
+                contractCreate(CONTRACT).maxAutomaticTokenAssociations(1),
+                getContractInfo(CONTRACT)
+                        .has(ContractInfoAsserts.contractWith().maxAutoAssociations(1))
+                        .logged(),
+                withOpContext((spec, opLog) -> {
+                    final var token = spec.registry().getTokenID(NFT_TOKEN);
+                    final var sender = spec.registry().getAccountID(SENDER);
+                    final var receiver = spec.registry().getAccountID(RECEIVER);
+                    final var amountToBeSent = 50L;
 
-                            allRunFor(
-                                    spec,
-                                    newKeyNamed(DELEGATE_KEY)
-                                            .shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
-                                    cryptoUpdate(SENDER).key(DELEGATE_KEY),
-                                    cryptoUpdate(RECEIVER).key(DELEGATE_KEY),
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(EMPTY_TUPLE_ARRAY)
-                                                            .build(),
-                                                    wrapIntoTupleArray(tokenTransferList()
-                                                            .forToken(token)
-                                                            .withNftTransfers(nftTransfer(sender, receiver, 1L, false))
-                                                            .build()))
-                                            .payingWith(SENDER)
-                                            .via(cryptoTransferTxnForNonFungible)
-                                            .gas(GAS_TO_OFFER));
-                        }),
-                        getTxnRecord(cryptoTransferTxnForNonFungible)
-                                .andAllChildRecords()
-                                .logged())
-                .then(
-                        getTokenInfo(NFT_TOKEN).hasTotalSupply(2),
-                        getAccountInfo(RECEIVER).hasOwnedNfts(1),
-                        getAccountBalance(RECEIVER).hasTokenBalance(NFT_TOKEN, 1),
-                        getAccountInfo(SENDER).hasOwnedNfts(0),
-                        getAccountBalance(SENDER).hasTokenBalance(NFT_TOKEN, 0),
-                        getTokenInfo(NFT_TOKEN).logged(),
-                        childRecordsCheck(
-                                cryptoTransferTxnForNonFungible,
-                                SUCCESS,
-                                recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(SUCCESS)))
-                                        .tokenTransfers(NonFungibleTransfers.changingNFTBalances()
-                                                .including(NFT_TOKEN, SENDER, RECEIVER, 1L))
-                                        .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
-                                                .including(FEE_TOKEN, RECEIVER, -1)
-                                                .including(FEE_TOKEN, TOKEN_TREASURY, 1))
-                                        .assessedCustomFeeCount(1)));
+                    allRunFor(
+                            spec,
+                            newKeyNamed(DELEGATE_KEY).shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
+                            cryptoUpdate(SENDER).key(DELEGATE_KEY),
+                            cryptoUpdate(RECEIVER).key(DELEGATE_KEY),
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(EMPTY_TUPLE_ARRAY)
+                                                    .build(),
+                                            wrapIntoTupleArray(tokenTransferList()
+                                                    .forToken(token)
+                                                    .withNftTransfers(nftTransfer(sender, receiver, 1L, false))
+                                                    .build()))
+                                    .payingWith(SENDER)
+                                    .via(cryptoTransferTxnForNonFungible)
+                                    .gas(GAS_TO_OFFER));
+                }),
+                getTxnRecord(cryptoTransferTxnForNonFungible)
+                        .andAllChildRecords()
+                        .logged(),
+                getTokenInfo(NFT_TOKEN).hasTotalSupply(2),
+                getAccountInfo(RECEIVER).hasOwnedNfts(1),
+                getAccountBalance(RECEIVER).hasTokenBalance(NFT_TOKEN, 1),
+                getAccountInfo(SENDER).hasOwnedNfts(0),
+                getAccountBalance(SENDER).hasTokenBalance(NFT_TOKEN, 0),
+                getTokenInfo(NFT_TOKEN).logged(),
+                childRecordsCheck(
+                        cryptoTransferTxnForNonFungible,
+                        SUCCESS,
+                        recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SUCCESS)))
+                                .tokenTransfers(NonFungibleTransfers.changingNFTBalances()
+                                        .including(NFT_TOKEN, SENDER, RECEIVER, 1L))
+                                .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
+                                        .including(FEE_TOKEN, RECEIVER, -1)
+                                        .including(FEE_TOKEN, TOKEN_TREASURY, 1))
+                                .assessedCustomFeeCount(1)));
     }
 
     @HapiTest
@@ -605,236 +548,202 @@ public class AtomicCryptoTransferHTSSuite {
         final var cryptoTransferTxnForNft = "cryptoTransferTxnForNft";
         final var cryptoTransferRevertNoKeyTxn = "cryptoTransferRevertNoKeyTxn";
 
-        return propertyPreservingHapiSpec(
-                        "cryptoTransferForNonFungibleTokenOnly",
-                        NONDETERMINISTIC_TRANSACTION_FEES,
-                        NONDETERMINISTIC_FUNCTION_PARAMETERS,
-                        NONDETERMINISTIC_NONCE)
-                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
-                .given(
-                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
-                        newKeyNamed(MULTI_KEY),
-                        cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(SENDER2).balance(10 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(RECEIVER).receiverSigRequired(true),
-                        cryptoCreate(TOKEN_TREASURY),
-                        tokenCreate(NFT_TOKEN)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .adminKey(MULTI_KEY)
-                                .supplyKey(MULTI_KEY)
-                                .supplyType(TokenSupplyType.INFINITE)
-                                .initialSupply(0)
-                                .treasury(TOKEN_TREASURY),
-                        tokenAssociate(SENDER, List.of(NFT_TOKEN)),
-                        tokenAssociate(SENDER2, List.of(NFT_TOKEN)),
-                        mintToken(NFT_TOKEN, List.of(metadata("firstMemo"), metadata("secondMemo"))),
-                        tokenAssociate(RECEIVER, List.of(NFT_TOKEN)),
-                        cryptoTransfer(TokenMovement.movingUnique(NFT_TOKEN, 1).between(TOKEN_TREASURY, SENDER))
-                                .payingWith(SENDER),
-                        uploadInitCode(CONTRACT),
-                        contractCreate(CONTRACT))
-                .when(
-                        withOpContext((spec, opLog) -> {
-                            final var token = spec.registry().getTokenID(NFT_TOKEN);
-                            final var sender = spec.registry().getAccountID(SENDER);
-                            final var sender2 = spec.registry().getAccountID(SENDER2);
-                            final var receiver = spec.registry().getAccountID(RECEIVER);
+        return hapiTest(
+                newKeyNamed(MULTI_KEY),
+                cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
+                cryptoCreate(SENDER2).balance(10 * ONE_HUNDRED_HBARS),
+                cryptoCreate(RECEIVER).receiverSigRequired(true),
+                cryptoCreate(TOKEN_TREASURY),
+                tokenCreate(NFT_TOKEN)
+                        .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
+                        .adminKey(MULTI_KEY)
+                        .supplyKey(MULTI_KEY)
+                        .supplyType(TokenSupplyType.INFINITE)
+                        .initialSupply(0)
+                        .treasury(TOKEN_TREASURY),
+                tokenAssociate(SENDER, List.of(NFT_TOKEN)),
+                tokenAssociate(SENDER2, List.of(NFT_TOKEN)),
+                mintToken(NFT_TOKEN, List.of(metadata("firstMemo"), metadata("secondMemo"))),
+                tokenAssociate(RECEIVER, List.of(NFT_TOKEN)),
+                cryptoTransfer(TokenMovement.movingUnique(NFT_TOKEN, 1).between(TOKEN_TREASURY, SENDER))
+                        .payingWith(SENDER),
+                uploadInitCode(CONTRACT),
+                contractCreate(CONTRACT),
+                withOpContext((spec, opLog) -> {
+                    final var token = spec.registry().getTokenID(NFT_TOKEN);
+                    final var sender = spec.registry().getAccountID(SENDER);
+                    final var sender2 = spec.registry().getAccountID(SENDER2);
+                    final var receiver = spec.registry().getAccountID(RECEIVER);
 
-                            allRunFor(
-                                    spec,
-                                    newKeyNamed(DELEGATE_KEY)
-                                            .shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
-                                    cryptoUpdate(SENDER).key(DELEGATE_KEY),
-                                    cryptoUpdate(RECEIVER).key(DELEGATE_KEY),
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(EMPTY_TUPLE_ARRAY)
-                                                            .build(),
-                                                    wrapIntoTupleArray(tokenTransferList()
-                                                            .forToken(token)
-                                                            .withNftTransfers(nftTransfer(sender, receiver, 1L, false))
-                                                            .build()))
-                                            .payingWith(GENESIS)
-                                            .via(cryptoTransferTxnForNft)
-                                            .gas(GAS_TO_OFFER),
-                                    // Ensure that the transfer fails when the sender does not have the correct key
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(EMPTY_TUPLE_ARRAY)
-                                                            .build(),
-                                                    wrapIntoTupleArray(tokenTransferList()
-                                                            .forToken(token)
-                                                            .withNftTransfers(nftTransfer(sender2, receiver, 1L, false))
-                                                            .build()))
-                                            .payingWith(GENESIS)
-                                            .via(cryptoTransferRevertNoKeyTxn)
-                                            .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
-                                            .gas(GAS_TO_OFFER));
-                        }),
-                        getTxnRecord(cryptoTransferTxnForNft)
-                                .andAllChildRecords()
-                                .logged())
-                .then(
-                        getTokenInfo(NFT_TOKEN).hasTotalSupply(2),
-                        getAccountInfo(RECEIVER).hasOwnedNfts(1),
-                        getAccountBalance(RECEIVER).hasTokenBalance(NFT_TOKEN, 1),
-                        getAccountInfo(SENDER).hasOwnedNfts(0),
-                        getAccountBalance(SENDER).hasTokenBalance(NFT_TOKEN, 0),
-                        getTokenInfo(NFT_TOKEN).logged(),
-                        childRecordsCheck(
-                                cryptoTransferTxnForNft,
-                                SUCCESS,
-                                recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(SUCCESS)))
-                                        .tokenTransfers(NonFungibleTransfers.changingNFTBalances()
-                                                .including(NFT_TOKEN, SENDER, RECEIVER, 1L))),
-                        childRecordsCheck(
-                                cryptoTransferRevertNoKeyTxn,
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith()
-                                        .status(SPENDER_DOES_NOT_HAVE_ALLOWANCE)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(htsPrecompileResult()
-                                                        .withStatus(SPENDER_DOES_NOT_HAVE_ALLOWANCE)))));
+                    allRunFor(
+                            spec,
+                            newKeyNamed(DELEGATE_KEY).shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
+                            cryptoUpdate(SENDER).key(DELEGATE_KEY),
+                            cryptoUpdate(RECEIVER).key(DELEGATE_KEY),
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(EMPTY_TUPLE_ARRAY)
+                                                    .build(),
+                                            wrapIntoTupleArray(tokenTransferList()
+                                                    .forToken(token)
+                                                    .withNftTransfers(nftTransfer(sender, receiver, 1L, false))
+                                                    .build()))
+                                    .payingWith(GENESIS)
+                                    .via(cryptoTransferTxnForNft)
+                                    .gas(GAS_TO_OFFER),
+                            // Ensure that the transfer fails when the sender does not have the correct key
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(EMPTY_TUPLE_ARRAY)
+                                                    .build(),
+                                            wrapIntoTupleArray(tokenTransferList()
+                                                    .forToken(token)
+                                                    .withNftTransfers(nftTransfer(sender2, receiver, 1L, false))
+                                                    .build()))
+                                    .payingWith(GENESIS)
+                                    .via(cryptoTransferRevertNoKeyTxn)
+                                    .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
+                                    .gas(GAS_TO_OFFER));
+                }),
+                getTxnRecord(cryptoTransferTxnForNft).andAllChildRecords().logged(),
+                getTokenInfo(NFT_TOKEN).hasTotalSupply(2),
+                getAccountInfo(RECEIVER).hasOwnedNfts(1),
+                getAccountBalance(RECEIVER).hasTokenBalance(NFT_TOKEN, 1),
+                getAccountInfo(SENDER).hasOwnedNfts(0),
+                getAccountBalance(SENDER).hasTokenBalance(NFT_TOKEN, 0),
+                getTokenInfo(NFT_TOKEN).logged(),
+                childRecordsCheck(
+                        cryptoTransferTxnForNft,
+                        SUCCESS,
+                        recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SUCCESS)))
+                                .tokenTransfers(NonFungibleTransfers.changingNFTBalances()
+                                        .including(NFT_TOKEN, SENDER, RECEIVER, 1L))),
+                childRecordsCheck(
+                        cryptoTransferRevertNoKeyTxn,
+                        CONTRACT_REVERT_EXECUTED,
+                        recordWith()
+                                .status(SPENDER_DOES_NOT_HAVE_ALLOWANCE)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SPENDER_DOES_NOT_HAVE_ALLOWANCE)))));
     }
 
     @HapiTest
     final Stream<DynamicTest> cryptoTransferHBarFungibleNft() {
         final var cryptoTransferTxnForAll = "cryptoTransferTxnForAll";
 
-        return propertyPreservingHapiSpec(
-                        "cryptoTransferHBarFungibleNft",
-                        NONDETERMINISTIC_FUNCTION_PARAMETERS,
-                        NONDETERMINISTIC_TRANSACTION_FEES,
-                        ACCEPTED_MONO_GAS_CALCULATION_DIFFERENCE)
-                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
-                .given(
-                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
-                        newKeyNamed(MULTI_KEY),
-                        cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
-                        cryptoCreate(SENDER2).balance(5 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(RECEIVER2).receiverSigRequired(true),
-                        cryptoCreate(TOKEN_TREASURY),
-                        tokenCreate(FUNGIBLE_TOKEN)
-                                .tokenType(TokenType.FUNGIBLE_COMMON)
-                                .initialSupply(TOTAL_SUPPLY)
-                                .treasury(TOKEN_TREASURY),
-                        tokenCreate(NFT_TOKEN)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .adminKey(MULTI_KEY)
-                                .supplyKey(MULTI_KEY)
-                                .supplyType(TokenSupplyType.INFINITE)
-                                .initialSupply(0)
-                                .treasury(TOKEN_TREASURY),
-                        tokenAssociate(SENDER, List.of(FUNGIBLE_TOKEN)),
-                        tokenAssociate(SENDER2, List.of(NFT_TOKEN)),
-                        mintToken(NFT_TOKEN, List.of(metadata("firstMemo"), metadata("secondMemo"))),
-                        tokenAssociate(RECEIVER, List.of(FUNGIBLE_TOKEN)),
-                        tokenAssociate(RECEIVER2, List.of(NFT_TOKEN)),
-                        cryptoTransfer(moving(200, FUNGIBLE_TOKEN).between(TOKEN_TREASURY, SENDER))
-                                .payingWith(GENESIS),
-                        cryptoTransfer(TokenMovement.movingUnique(NFT_TOKEN, 1).between(TOKEN_TREASURY, SENDER2))
-                                .payingWith(SENDER2),
-                        uploadInitCode(CONTRACT),
-                        contractCreate(CONTRACT))
-                .when(
-                        withOpContext((spec, opLog) -> {
-                            final var fungibleToken = spec.registry().getTokenID(FUNGIBLE_TOKEN);
-                            final var nonFungibleToken = spec.registry().getTokenID(NFT_TOKEN);
-                            final var fungibleTokenSender = spec.registry().getAccountID(SENDER);
-                            final var fungibleTokenReceiver = spec.registry().getAccountID(RECEIVER);
-                            final var nonFungibleTokenSender = spec.registry().getAccountID(SENDER2);
-                            final var nonFungibleTokenReceiver = spec.registry().getAccountID(RECEIVER2);
-                            final var amountToBeSent = 50 * ONE_HBAR;
+        return hapiTest(
+                newKeyNamed(MULTI_KEY),
+                cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
+                cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
+                cryptoCreate(SENDER2).balance(5 * ONE_HUNDRED_HBARS),
+                cryptoCreate(RECEIVER2).receiverSigRequired(true),
+                cryptoCreate(TOKEN_TREASURY),
+                tokenCreate(FUNGIBLE_TOKEN)
+                        .tokenType(TokenType.FUNGIBLE_COMMON)
+                        .initialSupply(TOTAL_SUPPLY)
+                        .treasury(TOKEN_TREASURY),
+                tokenCreate(NFT_TOKEN)
+                        .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
+                        .adminKey(MULTI_KEY)
+                        .supplyKey(MULTI_KEY)
+                        .supplyType(TokenSupplyType.INFINITE)
+                        .initialSupply(0)
+                        .treasury(TOKEN_TREASURY),
+                tokenAssociate(SENDER, List.of(FUNGIBLE_TOKEN)),
+                tokenAssociate(SENDER2, List.of(NFT_TOKEN)),
+                mintToken(NFT_TOKEN, List.of(metadata("firstMemo"), metadata("secondMemo"))),
+                tokenAssociate(RECEIVER, List.of(FUNGIBLE_TOKEN)),
+                tokenAssociate(RECEIVER2, List.of(NFT_TOKEN)),
+                cryptoTransfer(moving(200, FUNGIBLE_TOKEN).between(TOKEN_TREASURY, SENDER))
+                        .payingWith(GENESIS),
+                cryptoTransfer(TokenMovement.movingUnique(NFT_TOKEN, 1).between(TOKEN_TREASURY, SENDER2))
+                        .payingWith(SENDER2),
+                uploadInitCode(CONTRACT),
+                contractCreate(CONTRACT),
+                withOpContext((spec, opLog) -> {
+                    final var fungibleToken = spec.registry().getTokenID(FUNGIBLE_TOKEN);
+                    final var nonFungibleToken = spec.registry().getTokenID(NFT_TOKEN);
+                    final var fungibleTokenSender = spec.registry().getAccountID(SENDER);
+                    final var fungibleTokenReceiver = spec.registry().getAccountID(RECEIVER);
+                    final var nonFungibleTokenSender = spec.registry().getAccountID(SENDER2);
+                    final var nonFungibleTokenReceiver = spec.registry().getAccountID(RECEIVER2);
+                    final var amountToBeSent = 50 * ONE_HBAR;
 
-                            allRunFor(
-                                    spec,
-                                    newKeyNamed(DELEGATE_KEY)
-                                            .shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
-                                    cryptoUpdate(SENDER).key(DELEGATE_KEY),
-                                    cryptoUpdate(SENDER2).key(DELEGATE_KEY),
-                                    cryptoUpdate(RECEIVER).key(DELEGATE_KEY),
-                                    cryptoUpdate(RECEIVER2).key(DELEGATE_KEY),
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(
-                                                                    accountAmount(
-                                                                            fungibleTokenSender,
-                                                                            -amountToBeSent,
-                                                                            false),
-                                                                    accountAmount(
-                                                                            fungibleTokenReceiver,
-                                                                            amountToBeSent,
-                                                                            false))
-                                                            .build(),
-                                                    tokenTransferLists()
-                                                            .withTokenTransferList(
-                                                                    tokenTransferList()
-                                                                            .forToken(fungibleToken)
-                                                                            .withAccountAmounts(
-                                                                                    accountAmount(
-                                                                                            fungibleTokenSender,
-                                                                                            -45L,
-                                                                                            false),
-                                                                                    accountAmount(
-                                                                                            fungibleTokenReceiver,
-                                                                                            45L,
-                                                                                            false))
-                                                                            .build(),
-                                                                    tokenTransferList()
-                                                                            .forToken(nonFungibleToken)
-                                                                            .withNftTransfers(
-                                                                                    nftTransfer(
-                                                                                            nonFungibleTokenSender,
-                                                                                            nonFungibleTokenReceiver,
-                                                                                            1L,
-                                                                                            false))
-                                                                            .build())
-                                                            .build())
-                                            .payingWith(GENESIS)
-                                            .via(cryptoTransferTxnForAll)
-                                            .gas(GAS_TO_OFFER));
-                        }),
-                        getTxnRecord(cryptoTransferTxnForAll)
-                                .andAllChildRecords()
-                                .logged())
-                .then(
-                        getTokenInfo(FUNGIBLE_TOKEN).hasTotalSupply(TOTAL_SUPPLY),
-                        getAccountBalance(RECEIVER).hasTokenBalance(FUNGIBLE_TOKEN, 45),
-                        getAccountBalance(SENDER).hasTokenBalance(FUNGIBLE_TOKEN, 155),
-                        getTokenInfo(FUNGIBLE_TOKEN).logged(),
-                        getTokenInfo(NFT_TOKEN).hasTotalSupply(2),
-                        getAccountInfo(RECEIVER2).hasOwnedNfts(1),
-                        getAccountBalance(RECEIVER2).hasTokenBalance(NFT_TOKEN, 1),
-                        getAccountInfo(SENDER2).hasOwnedNfts(0),
-                        getAccountBalance(SENDER2).hasTokenBalance(NFT_TOKEN, 0),
-                        getTokenInfo(NFT_TOKEN).logged(),
-                        getAccountBalance(SENDER).hasTinyBars(950 * ONE_HBAR),
-                        getAccountBalance(RECEIVER).hasTinyBars(250 * ONE_HBAR),
-                        childRecordsCheck(
-                                cryptoTransferTxnForAll,
-                                SUCCESS,
-                                recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(SUCCESS)))
-                                        .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
-                                                .including(FUNGIBLE_TOKEN, SENDER, -45L)
-                                                .including(FUNGIBLE_TOKEN, RECEIVER, 45L))
-                                        .tokenTransfers(NonFungibleTransfers.changingNFTBalances()
-                                                .including(NFT_TOKEN, SENDER2, RECEIVER2, 1L))
-                                        .transfers(including(tinyBarsFromTo(SENDER, RECEIVER, 50 * ONE_HBAR)))));
+                    allRunFor(
+                            spec,
+                            newKeyNamed(DELEGATE_KEY).shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
+                            cryptoUpdate(SENDER).key(DELEGATE_KEY),
+                            cryptoUpdate(SENDER2).key(DELEGATE_KEY),
+                            cryptoUpdate(RECEIVER).key(DELEGATE_KEY),
+                            cryptoUpdate(RECEIVER2).key(DELEGATE_KEY),
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(
+                                                            accountAmount(fungibleTokenSender, -amountToBeSent, false),
+                                                            accountAmount(fungibleTokenReceiver, amountToBeSent, false))
+                                                    .build(),
+                                            tokenTransferLists()
+                                                    .withTokenTransferList(
+                                                            tokenTransferList()
+                                                                    .forToken(fungibleToken)
+                                                                    .withAccountAmounts(
+                                                                            accountAmount(
+                                                                                    fungibleTokenSender, -45L, false),
+                                                                            accountAmount(
+                                                                                    fungibleTokenReceiver, 45L, false))
+                                                                    .build(),
+                                                            tokenTransferList()
+                                                                    .forToken(nonFungibleToken)
+                                                                    .withNftTransfers(
+                                                                            nftTransfer(
+                                                                                    nonFungibleTokenSender,
+                                                                                    nonFungibleTokenReceiver,
+                                                                                    1L,
+                                                                                    false))
+                                                                    .build())
+                                                    .build())
+                                    .payingWith(GENESIS)
+                                    .via(cryptoTransferTxnForAll)
+                                    .gas(GAS_TO_OFFER));
+                }),
+                getTxnRecord(cryptoTransferTxnForAll).andAllChildRecords().logged(),
+                getTokenInfo(FUNGIBLE_TOKEN).hasTotalSupply(TOTAL_SUPPLY),
+                getAccountBalance(RECEIVER).hasTokenBalance(FUNGIBLE_TOKEN, 45),
+                getAccountBalance(SENDER).hasTokenBalance(FUNGIBLE_TOKEN, 155),
+                getTokenInfo(FUNGIBLE_TOKEN).logged(),
+                getTokenInfo(NFT_TOKEN).hasTotalSupply(2),
+                getAccountInfo(RECEIVER2).hasOwnedNfts(1),
+                getAccountBalance(RECEIVER2).hasTokenBalance(NFT_TOKEN, 1),
+                getAccountInfo(SENDER2).hasOwnedNfts(0),
+                getAccountBalance(SENDER2).hasTokenBalance(NFT_TOKEN, 0),
+                getTokenInfo(NFT_TOKEN).logged(),
+                getAccountBalance(SENDER).hasTinyBars(950 * ONE_HBAR),
+                getAccountBalance(RECEIVER).hasTinyBars(250 * ONE_HBAR),
+                childRecordsCheck(
+                        cryptoTransferTxnForAll,
+                        SUCCESS,
+                        recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SUCCESS)))
+                                .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
+                                        .including(FUNGIBLE_TOKEN, SENDER, -45L)
+                                        .including(FUNGIBLE_TOKEN, RECEIVER, 45L))
+                                .tokenTransfers(NonFungibleTransfers.changingNFTBalances()
+                                        .including(NFT_TOKEN, SENDER2, RECEIVER2, 1L))
+                                .transfers(including(tinyBarsFromTo(SENDER, RECEIVER, 50 * ONE_HBAR)))));
     }
 
     @HapiTest
@@ -846,159 +755,154 @@ public class AtomicCryptoTransferHTSSuite {
         final var revertingTransferFromTxn = "revertWhenMoreThanAllowance";
         final var revertingTransferFromTxn2 = "revertingTxn";
 
-        return propertyPreservingHapiSpec("cryptoTransferAllowanceToContractHbar")
-                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
-                .given(
-                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
-                        newKeyNamed(MULTI_KEY),
-                        cryptoCreate(OWNER).balance(100 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(RECEIVER),
-                        uploadInitCode(CONTRACT),
-                        contractCreate(CONTRACT))
-                .when(
-                        cryptoApproveAllowance()
-                                .payingWith(DEFAULT_PAYER)
-                                .addCryptoAllowance(OWNER, CONTRACT, allowance)
-                                .via(BASE_APPROVAL_TXN)
-                                .signedBy(DEFAULT_PAYER, OWNER)
-                                .logged(),
-                        getAccountDetails(OWNER)
-                                .payingWith(GENESIS)
-                                .has(accountDetailsWith().cryptoAllowancesContaining(CONTRACT, allowance))
-                                .logged(),
-                        withOpContext((spec, opLog) -> {
-                            final var owner = spec.registry().getAccountID(OWNER);
-                            final var receiver = spec.registry().getAccountID(RECEIVER);
+        return hapiTest(
+                newKeyNamed(MULTI_KEY),
+                cryptoCreate(OWNER).balance(100 * ONE_HUNDRED_HBARS),
+                cryptoCreate(RECEIVER),
+                uploadInitCode(CONTRACT),
+                contractCreate(CONTRACT),
+                cryptoApproveAllowance()
+                        .payingWith(DEFAULT_PAYER)
+                        .addCryptoAllowance(OWNER, CONTRACT, allowance)
+                        .via(BASE_APPROVAL_TXN)
+                        .signedBy(DEFAULT_PAYER, OWNER)
+                        .logged(),
+                getAccountDetails(OWNER)
+                        .payingWith(GENESIS)
+                        .has(accountDetailsWith().cryptoAllowancesContaining(CONTRACT, allowance))
+                        .logged(),
+                withOpContext((spec, opLog) -> {
+                    final var owner = spec.registry().getAccountID(OWNER);
+                    final var receiver = spec.registry().getAccountID(RECEIVER);
 
-                            /*
-                             We will be covering the following test cases.  These test cover hbar transfers
-                             1. Transfer more than allowance amount
-                             2. Transfer when there is no approval - this now succeeds as the transaction is retried with the allowance
-                             3. Transfer 1/2 the allowance amount
-                             4. Transfer the other 1/2 of the allowance amount
-                             5. Transfer after allowance is spent
-                            */
+                    /*
+                     We will be covering the following test cases.  These test cover hbar transfers
+                     1. Transfer more than allowance amount
+                     2. Transfer when there is no approval - this now succeeds as the transaction is retried with the allowance
+                     3. Transfer 1/2 the allowance amount
+                     4. Transfer the other 1/2 of the allowance amount
+                     5. Transfer after allowance is spent
+                    */
 
-                            allRunFor(
-                                    spec,
-                                    // Try to send 1 more than the allowance amount from
-                                    // owner to receiver
-                                    // should fail
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(
-                                                                    accountAmount(owner, -(allowance + 1), true),
-                                                                    accountAmount(receiver, allowance + 1, true))
-                                                            .build(),
-                                                    EMPTY_TUPLE_ARRAY)
-                                            .via(revertingTransferFromTxn)
-                                            .hasKnownStatus(CONTRACT_REVERT_EXECUTED),
-                                    // Try to send allowance amount but turn off isApproval
-                                    // flag
-                                    // This used to fail but now succeeds as the transaction is
-                                    // automatically retried as with the allowance
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(
-                                                                    accountAmount(owner, -1L, false),
-                                                                    accountAmount(receiver, 1L, false))
-                                                            .build(),
-                                                    EMPTY_TUPLE_ARRAY)
-                                            .via(successfulTransferFromTxn)
-                                            .hasKnownStatus(SUCCESS),
-                                    // Try to send 1/2 of the allowance amount from owner to
-                                    // receiver
-                                    // should succeed as isApproval is true.
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(
-                                                                    accountAmount(owner, -(allowance - 1) / 2, true),
-                                                                    accountAmount(receiver, (allowance - 1) / 2, true))
-                                                            .build(),
-                                                    EMPTY_TUPLE_ARRAY)
-                                            .via(successfulTransferFromTxn2)
-                                            .hasKnownStatus(SUCCESS),
-                                    // Try to send second 1/2 of the allowance amount from
-                                    // owner to receiver
-                                    // should succeed as isApproval is true.
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(
-                                                                    accountAmount(owner, -(allowance - 1) / 2, true),
-                                                                    accountAmount(receiver, (allowance - 1) / 2, true))
-                                                            .build(),
-                                                    EMPTY_TUPLE_ARRAY)
-                                            .via(successfulTransferFromTxn3)
-                                            .hasKnownStatus(SUCCESS),
-                                    getAccountDetails(OWNER)
-                                            .payingWith(GENESIS)
-                                            .has(accountDetailsWith().noAllowances()),
-                                    // Try to send 1 hbar from owner to receiver
-                                    // should fail as all allowance has been spent
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(
-                                                                    accountAmount(owner, -1L, true),
-                                                                    accountAmount(receiver, 1L, true))
-                                                            .build(),
-                                                    EMPTY_TUPLE_ARRAY)
-                                            .via(revertingTransferFromTxn2)
-                                            .hasKnownStatus(CONTRACT_REVERT_EXECUTED));
-                        }))
-                .then(
-                        childRecordsCheck(
-                                revertingTransferFromTxn,
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith()
-                                        .status(AMOUNT_EXCEEDS_ALLOWANCE)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(AMOUNT_EXCEEDS_ALLOWANCE)))),
-                        childRecordsCheck(
-                                successfulTransferFromTxn,
-                                SUCCESS,
-                                recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(SUCCESS)))
-                                        .transfers(including(tinyBarsFromTo(OWNER, RECEIVER, 1)))),
-                        childRecordsCheck(
-                                successfulTransferFromTxn2,
-                                SUCCESS,
-                                recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(SUCCESS)))
-                                        .transfers(including(tinyBarsFromTo(OWNER, RECEIVER, (allowance - 1) / 2)))),
-                        childRecordsCheck(
-                                successfulTransferFromTxn3,
-                                SUCCESS,
-                                recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(SUCCESS)))
-                                        .transfers(including(tinyBarsFromTo(OWNER, RECEIVER, (allowance - 1) / 2)))),
-                        childRecordsCheck(
-                                revertingTransferFromTxn2,
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith()
-                                        .status(SPENDER_DOES_NOT_HAVE_ALLOWANCE)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(htsPrecompileResult()
-                                                        .withStatus(SPENDER_DOES_NOT_HAVE_ALLOWANCE)))));
+                    allRunFor(
+                            spec,
+                            // Try to send 1 more than the allowance amount from
+                            // owner to receiver
+                            // should fail
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(
+                                                            accountAmount(owner, -(allowance + 1), true),
+                                                            accountAmount(receiver, allowance + 1, true))
+                                                    .build(),
+                                            EMPTY_TUPLE_ARRAY)
+                                    .via(revertingTransferFromTxn)
+                                    .hasKnownStatus(CONTRACT_REVERT_EXECUTED),
+                            // Try to send allowance amount but turn off isApproval
+                            // flag
+                            // This used to fail but now succeeds as the transaction is
+                            // automatically retried as with the allowance
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(
+                                                            accountAmount(owner, -1L, false),
+                                                            accountAmount(receiver, 1L, false))
+                                                    .build(),
+                                            EMPTY_TUPLE_ARRAY)
+                                    .via(successfulTransferFromTxn)
+                                    .hasKnownStatus(SUCCESS),
+                            // Try to send 1/2 of the allowance amount from owner to
+                            // receiver
+                            // should succeed as isApproval is true.
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(
+                                                            accountAmount(owner, -(allowance - 1) / 2, true),
+                                                            accountAmount(receiver, (allowance - 1) / 2, true))
+                                                    .build(),
+                                            EMPTY_TUPLE_ARRAY)
+                                    .via(successfulTransferFromTxn2)
+                                    .hasKnownStatus(SUCCESS),
+                            // Try to send second 1/2 of the allowance amount from
+                            // owner to receiver
+                            // should succeed as isApproval is true.
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(
+                                                            accountAmount(owner, -(allowance - 1) / 2, true),
+                                                            accountAmount(receiver, (allowance - 1) / 2, true))
+                                                    .build(),
+                                            EMPTY_TUPLE_ARRAY)
+                                    .via(successfulTransferFromTxn3)
+                                    .hasKnownStatus(SUCCESS),
+                            getAccountDetails(OWNER)
+                                    .payingWith(GENESIS)
+                                    .has(accountDetailsWith().noAllowances()),
+                            // Try to send 1 hbar from owner to receiver
+                            // should fail as all allowance has been spent
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(
+                                                            accountAmount(owner, -1L, true),
+                                                            accountAmount(receiver, 1L, true))
+                                                    .build(),
+                                            EMPTY_TUPLE_ARRAY)
+                                    .via(revertingTransferFromTxn2)
+                                    .hasKnownStatus(CONTRACT_REVERT_EXECUTED));
+                }),
+                childRecordsCheck(
+                        revertingTransferFromTxn,
+                        CONTRACT_REVERT_EXECUTED,
+                        recordWith()
+                                .status(AMOUNT_EXCEEDS_ALLOWANCE)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(AMOUNT_EXCEEDS_ALLOWANCE)))),
+                childRecordsCheck(
+                        successfulTransferFromTxn,
+                        SUCCESS,
+                        recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SUCCESS)))
+                                .transfers(including(tinyBarsFromTo(OWNER, RECEIVER, 1)))),
+                childRecordsCheck(
+                        successfulTransferFromTxn2,
+                        SUCCESS,
+                        recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SUCCESS)))
+                                .transfers(including(tinyBarsFromTo(OWNER, RECEIVER, (allowance - 1) / 2)))),
+                childRecordsCheck(
+                        successfulTransferFromTxn3,
+                        SUCCESS,
+                        recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SUCCESS)))
+                                .transfers(including(tinyBarsFromTo(OWNER, RECEIVER, (allowance - 1) / 2)))),
+                childRecordsCheck(
+                        revertingTransferFromTxn2,
+                        CONTRACT_REVERT_EXECUTED,
+                        recordWith()
+                                .status(SPENDER_DOES_NOT_HAVE_ALLOWANCE)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SPENDER_DOES_NOT_HAVE_ALLOWANCE)))));
     }
 
     @HapiTest
@@ -1009,33 +913,29 @@ public class AtomicCryptoTransferHTSSuite {
         final var successfulTransferFromTxn3 = "txn3";
         final var revertingTransferFromTxnFungible = "revertWhenMoreThanAllowanceFungible";
         final var revertingTransferFromTxn2 = "revertingTxn";
-        return propertyPreservingHapiSpec("cryptoTransferAllowanceToContractFT")
-                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
-                .given(
-                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
-                        newKeyNamed(MULTI_KEY),
-                        cryptoCreate(OWNER).balance(100 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(RECEIVER).maxAutomaticTokenAssociations(1),
-                        tokenCreate(FUNGIBLE_TOKEN)
-                                .tokenType(TokenType.FUNGIBLE_COMMON)
-                                .supplyType(TokenSupplyType.FINITE)
-                                .initialSupply(allowance)
-                                .maxSupply(1000L)
-                                .supplyKey(MULTI_KEY)
-                                .treasury(OWNER),
-                        uploadInitCode(CONTRACT),
-                        contractCreate(CONTRACT),
-                        cryptoApproveAllowance()
-                                .payingWith(DEFAULT_PAYER)
-                                .addTokenAllowance(OWNER, FUNGIBLE_TOKEN, CONTRACT, allowance)
-                                .via(BASE_APPROVAL_TXN)
-                                .signedBy(DEFAULT_PAYER, OWNER)
-                                .fee(ONE_HBAR),
-                        getAccountDetails(OWNER)
-                                .payingWith(GENESIS)
-                                .has(accountDetailsWith()
-                                        .tokenAllowancesContaining(FUNGIBLE_TOKEN, CONTRACT, allowance)))
-                .when(withOpContext((spec, opLog) -> {
+        return hapiTest(
+                newKeyNamed(MULTI_KEY),
+                cryptoCreate(OWNER).balance(100 * ONE_HUNDRED_HBARS),
+                cryptoCreate(RECEIVER).maxAutomaticTokenAssociations(1),
+                tokenCreate(FUNGIBLE_TOKEN)
+                        .tokenType(TokenType.FUNGIBLE_COMMON)
+                        .supplyType(TokenSupplyType.FINITE)
+                        .initialSupply(allowance)
+                        .maxSupply(1000L)
+                        .supplyKey(MULTI_KEY)
+                        .treasury(OWNER),
+                uploadInitCode(CONTRACT),
+                contractCreate(CONTRACT),
+                cryptoApproveAllowance()
+                        .payingWith(DEFAULT_PAYER)
+                        .addTokenAllowance(OWNER, FUNGIBLE_TOKEN, CONTRACT, allowance)
+                        .via(BASE_APPROVAL_TXN)
+                        .signedBy(DEFAULT_PAYER, OWNER)
+                        .fee(ONE_HBAR),
+                getAccountDetails(OWNER)
+                        .payingWith(GENESIS)
+                        .has(accountDetailsWith().tokenAllowancesContaining(FUNGIBLE_TOKEN, CONTRACT, allowance)),
+                withOpContext((spec, opLog) -> {
                     final var token = spec.registry().getTokenID(FUNGIBLE_TOKEN);
                     final var owner = spec.registry().getAccountID(OWNER);
                     final var receiver = spec.registry().getAccountID(RECEIVER);
@@ -1139,94 +1039,86 @@ public class AtomicCryptoTransferHTSSuite {
                                                     .build()))
                                     .via(revertingTransferFromTxn2)
                                     .hasKnownStatus(CONTRACT_REVERT_EXECUTED));
-                }))
-                .then(
-                        childRecordsCheck(
-                                revertingTransferFromTxnFungible,
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith()
-                                        .status(AMOUNT_EXCEEDS_ALLOWANCE)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(AMOUNT_EXCEEDS_ALLOWANCE)))),
-                        childRecordsCheck(
-                                successfulTransferFromTxn,
-                                SUCCESS,
-                                recordWith().status(SUCCESS),
-                                recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(SUCCESS)))
-                                        .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
-                                                .including(FUNGIBLE_TOKEN, OWNER, -1L)
-                                                .including(FUNGIBLE_TOKEN, RECEIVER, 1L))),
-                        childRecordsCheck(
-                                successfulTransferFromTxn2,
-                                SUCCESS,
-                                recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(SUCCESS)))
-                                        .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
-                                                .including(FUNGIBLE_TOKEN, OWNER, -(allowance - 1) / 2)
-                                                .including(FUNGIBLE_TOKEN, RECEIVER, (allowance - 1) / 2))),
-                        childRecordsCheck(
-                                successfulTransferFromTxn3,
-                                SUCCESS,
-                                recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(SUCCESS)))
-                                        .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
-                                                .including(FUNGIBLE_TOKEN, OWNER, -(allowance - 1) / 2)
-                                                .including(FUNGIBLE_TOKEN, RECEIVER, (allowance - 1) / 2))),
-                        childRecordsCheck(
-                                revertingTransferFromTxn2,
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith()
-                                        .status(SPENDER_DOES_NOT_HAVE_ALLOWANCE)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(htsPrecompileResult()
-                                                        .withStatus(SPENDER_DOES_NOT_HAVE_ALLOWANCE)))));
+                }),
+                childRecordsCheck(
+                        revertingTransferFromTxnFungible,
+                        CONTRACT_REVERT_EXECUTED,
+                        recordWith()
+                                .status(AMOUNT_EXCEEDS_ALLOWANCE)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(AMOUNT_EXCEEDS_ALLOWANCE)))),
+                childRecordsCheck(
+                        successfulTransferFromTxn,
+                        SUCCESS,
+                        recordWith().status(SUCCESS),
+                        recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SUCCESS)))
+                                .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
+                                        .including(FUNGIBLE_TOKEN, OWNER, -1L)
+                                        .including(FUNGIBLE_TOKEN, RECEIVER, 1L))),
+                childRecordsCheck(
+                        successfulTransferFromTxn2,
+                        SUCCESS,
+                        recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SUCCESS)))
+                                .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
+                                        .including(FUNGIBLE_TOKEN, OWNER, -(allowance - 1) / 2)
+                                        .including(FUNGIBLE_TOKEN, RECEIVER, (allowance - 1) / 2))),
+                childRecordsCheck(
+                        successfulTransferFromTxn3,
+                        SUCCESS,
+                        recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SUCCESS)))
+                                .tokenTransfers(SomeFungibleTransfers.changingFungibleBalances()
+                                        .including(FUNGIBLE_TOKEN, OWNER, -(allowance - 1) / 2)
+                                        .including(FUNGIBLE_TOKEN, RECEIVER, (allowance - 1) / 2))),
+                childRecordsCheck(
+                        revertingTransferFromTxn2,
+                        CONTRACT_REVERT_EXECUTED,
+                        recordWith()
+                                .status(SPENDER_DOES_NOT_HAVE_ALLOWANCE)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SPENDER_DOES_NOT_HAVE_ALLOWANCE)))));
     }
 
     @HapiTest
     final Stream<DynamicTest> cryptoTransferAllowanceToContractNFT() {
         final var successfulTransferFromTxn = "txn";
         final var revertingTransferFromTxnNft = "revertWhenMoreThanAllowanceNft";
-        return propertyPreservingHapiSpec(
-                        "cryptoTransferAllowanceToContractNFT",
-                        NONDETERMINISTIC_FUNCTION_PARAMETERS,
-                        ACCEPTED_MONO_GAS_CALCULATION_DIFFERENCE,
-                        NONDETERMINISTIC_TRANSACTION_FEES)
-                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
-                .given(
-                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
-                        newKeyNamed(MULTI_KEY),
-                        cryptoCreate(OWNER).balance(100 * ONE_HUNDRED_HBARS).maxAutomaticTokenAssociations(5),
-                        cryptoCreate(RECEIVER).maxAutomaticTokenAssociations(5),
-                        tokenCreate(NFT_TOKEN)
-                                .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
-                                .treasury(OWNER)
-                                .initialSupply(0L)
-                                .supplyKey(MULTI_KEY),
-                        uploadInitCode(CONTRACT),
-                        contractCreate(CONTRACT),
-                        mintToken(
-                                NFT_TOKEN,
-                                List.of(
-                                        ByteStringUtils.wrapUnsafely("meta1".getBytes()),
-                                        ByteStringUtils.wrapUnsafely("meta2".getBytes()))),
-                        cryptoApproveAllowance()
-                                .payingWith(DEFAULT_PAYER)
-                                .addNftAllowance(OWNER, NFT_TOKEN, CONTRACT, false, List.of(2L))
-                                .via(BASE_APPROVAL_TXN)
-                                .signedBy(DEFAULT_PAYER, OWNER)
-                                .fee(ONE_HBAR))
-                .when(withOpContext((spec, opLog) -> {
+        return hapiTest(
+                newKeyNamed(MULTI_KEY),
+                cryptoCreate(OWNER).balance(100 * ONE_HUNDRED_HBARS).maxAutomaticTokenAssociations(5),
+                cryptoCreate(RECEIVER).maxAutomaticTokenAssociations(5),
+                tokenCreate(NFT_TOKEN)
+                        .tokenType(TokenType.NON_FUNGIBLE_UNIQUE)
+                        .treasury(OWNER)
+                        .initialSupply(0L)
+                        .supplyKey(MULTI_KEY),
+                uploadInitCode(CONTRACT),
+                contractCreate(CONTRACT),
+                mintToken(
+                        NFT_TOKEN,
+                        List.of(
+                                ByteStringUtils.wrapUnsafely("meta1".getBytes()),
+                                ByteStringUtils.wrapUnsafely("meta2".getBytes()))),
+                cryptoApproveAllowance()
+                        .payingWith(DEFAULT_PAYER)
+                        .addNftAllowance(OWNER, NFT_TOKEN, CONTRACT, false, List.of(2L))
+                        .via(BASE_APPROVAL_TXN)
+                        .signedBy(DEFAULT_PAYER, OWNER)
+                        .fee(ONE_HBAR),
+                withOpContext((spec, opLog) -> {
                     final var token = spec.registry().getTokenID(NFT_TOKEN);
                     final var owner = spec.registry().getAccountID(OWNER);
                     final var receiver = spec.registry().getAccountID(RECEIVER);
@@ -1259,68 +1151,59 @@ public class AtomicCryptoTransferHTSSuite {
                                                     .build()))
                                     .via(successfulTransferFromTxn)
                                     .hasKnownStatus(SUCCESS));
-                }))
-                .then(
-                        childRecordsCheck(
-                                revertingTransferFromTxnNft,
-                                CONTRACT_REVERT_EXECUTED,
-                                recordWith()
-                                        .status(SPENDER_DOES_NOT_HAVE_ALLOWANCE)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(htsPrecompileResult()
-                                                        .withStatus(SPENDER_DOES_NOT_HAVE_ALLOWANCE)))),
-                        childRecordsCheck(
-                                successfulTransferFromTxn,
-                                SUCCESS,
-                                recordWith().status(SUCCESS),
-                                recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .contractCallResult(
-                                                        htsPrecompileResult().withStatus(SUCCESS)))
-                                        .tokenTransfers(NonFungibleTransfers.changingNFTBalances()
-                                                .including(NFT_TOKEN, OWNER, RECEIVER, 2L))));
+                }),
+                childRecordsCheck(
+                        revertingTransferFromTxnNft,
+                        CONTRACT_REVERT_EXECUTED,
+                        recordWith()
+                                .status(SPENDER_DOES_NOT_HAVE_ALLOWANCE)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SPENDER_DOES_NOT_HAVE_ALLOWANCE)))),
+                childRecordsCheck(
+                        successfulTransferFromTxn,
+                        SUCCESS,
+                        recordWith().status(SUCCESS),
+                        recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .contractCallResult(
+                                                htsPrecompileResult().withStatus(SUCCESS)))
+                                .tokenTransfers(NonFungibleTransfers.changingNFTBalances()
+                                        .including(NFT_TOKEN, OWNER, RECEIVER, 2L))));
     }
 
     @HapiTest
     final Stream<DynamicTest> cryptoTransferAllowanceToContractFromContract() {
         final var successfulTransferFromTxn = "txn";
-        final var revertingTransferFromTxnNft = "revertWhenMoreThanAllowanceNft";
         final var simpleStorageContract = "SimpleStorage";
         final long allowance = 10L;
 
-        return propertyPreservingHapiSpec(
-                        "cryptoTransferAllowanceToContractFromContract",
-                        NONDETERMINISTIC_FUNCTION_PARAMETERS,
-                        ACCEPTED_MONO_GAS_CALCULATION_DIFFERENCE,
-                        NONDETERMINISTIC_TRANSACTION_FEES)
-                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
-                .given(
-                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
-                        newKeyNamed(MULTI_KEY),
-                        cryptoCreate(OWNER).balance(100 * ONE_HUNDRED_HBARS).maxAutomaticTokenAssociations(5),
-                        cryptoCreate(RECEIVER).maxAutomaticTokenAssociations(5),
-                        uploadInitCode(simpleStorageContract),
-                        contractCreate(simpleStorageContract),
-                        tokenCreate(FUNGIBLE_TOKEN)
-                                .tokenType(TokenType.FUNGIBLE_COMMON)
-                                .supplyType(TokenSupplyType.FINITE)
-                                .initialSupply(1000L)
-                                .maxSupply(1000L)
-                                .supplyKey(MULTI_KEY)
-                                .treasury(OWNER),
-                        uploadInitCode(CONTRACT),
-                        contractCreate(CONTRACT),
-                        tokenAssociate(simpleStorageContract, List.of(FUNGIBLE_TOKEN)),
-                        cryptoTransfer(moving(100L, FUNGIBLE_TOKEN).between(OWNER, simpleStorageContract))
-                                .payingWith(OWNER),
-                        cryptoApproveAllowance()
-                                .payingWith(DEFAULT_PAYER)
-                                .addTokenAllowance(simpleStorageContract, FUNGIBLE_TOKEN, CONTRACT, allowance)
-                                .via(BASE_APPROVAL_TXN)
-                                .signedBy(DEFAULT_PAYER, simpleStorageContract)
-                                .fee(ONE_HBAR))
-                .when(withOpContext((spec, opLog) -> {
+        return hapiTest(
+                newKeyNamed(MULTI_KEY),
+                cryptoCreate(OWNER).balance(100 * ONE_HUNDRED_HBARS).maxAutomaticTokenAssociations(5),
+                cryptoCreate(RECEIVER).maxAutomaticTokenAssociations(5),
+                uploadInitCode(simpleStorageContract),
+                contractCreate(simpleStorageContract),
+                tokenCreate(FUNGIBLE_TOKEN)
+                        .tokenType(TokenType.FUNGIBLE_COMMON)
+                        .supplyType(TokenSupplyType.FINITE)
+                        .initialSupply(1000L)
+                        .maxSupply(1000L)
+                        .supplyKey(MULTI_KEY)
+                        .treasury(OWNER),
+                uploadInitCode(CONTRACT),
+                contractCreate(CONTRACT),
+                tokenAssociate(simpleStorageContract, List.of(FUNGIBLE_TOKEN)),
+                cryptoTransfer(moving(100L, FUNGIBLE_TOKEN).between(OWNER, simpleStorageContract))
+                        .payingWith(OWNER),
+                cryptoApproveAllowance()
+                        .payingWith(DEFAULT_PAYER)
+                        .addTokenAllowance(simpleStorageContract, FUNGIBLE_TOKEN, CONTRACT, allowance)
+                        .via(BASE_APPROVAL_TXN)
+                        .signedBy(DEFAULT_PAYER, simpleStorageContract)
+                        .fee(ONE_HBAR),
+                withOpContext((spec, opLog) -> {
                     final var token = spec.registry().getTokenID(FUNGIBLE_TOKEN);
                     final var owner = spec.registry().getAccountID(simpleStorageContract);
                     final var receiver = spec.registry().getAccountID(RECEIVER);
@@ -1341,8 +1224,7 @@ public class AtomicCryptoTransferHTSSuite {
                                                     .build()))
                                     .via(successfulTransferFromTxn)
                                     .hasKnownStatus(SUCCESS));
-                }))
-                .then();
+                }));
     }
 
     @HapiTest
@@ -1350,45 +1232,36 @@ public class AtomicCryptoTransferHTSSuite {
         final var failedTransferFromTxn = "failed_txn";
         final long allowance = 20L;
 
-        return propertyPreservingHapiSpec(
-                        "cryptoTransferAllowanceToEOA",
-                        NONDETERMINISTIC_FUNCTION_PARAMETERS,
-                        ACCEPTED_MONO_GAS_CALCULATION_DIFFERENCE,
-                        NONDETERMINISTIC_TRANSACTION_FEES)
-                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
-                .given(
-                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
-                        cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
-                        uploadInitCode(CONTRACT),
-                        contractCreate(CONTRACT))
-                .when(
-                        withOpContext((spec, opLog) -> {
-                            final var sender = spec.registry().getAccountID(SENDER);
-                            final var receiver = spec.registry().getAccountID(RECEIVER);
-                            final var amountToBeSent = 50 * ONE_HBAR;
+        return hapiTest(
+                cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
+                cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS).receiverSigRequired(true),
+                uploadInitCode(CONTRACT),
+                contractCreate(CONTRACT),
+                withOpContext((spec, opLog) -> {
+                    final var sender = spec.registry().getAccountID(SENDER);
+                    final var receiver = spec.registry().getAccountID(RECEIVER);
+                    final var amountToBeSent = 50 * ONE_HBAR;
 
-                            allRunFor(
-                                    spec,
-                                    newKeyNamed(DELEGATE_KEY)
-                                            .shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
-                                    cryptoUpdate(SENDER).key(DELEGATE_KEY),
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(
-                                                                    accountAmount(sender, -amountToBeSent, false),
-                                                                    accountAmount(receiver, amountToBeSent, false))
-                                                            .build(),
-                                                    EMPTY_TUPLE_ARRAY)
-                                            .payingWith(GENESIS)
-                                            .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
-                                            .via(failedTransferFromTxn)
-                                            .gas(GAS_TO_OFFER));
-                        }),
-                        getTxnRecord(failedTransferFromTxn).andAllChildRecords().logged())
-                .then(childRecordsCheck(
+                    allRunFor(
+                            spec,
+                            newKeyNamed(DELEGATE_KEY).shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
+                            cryptoUpdate(SENDER).key(DELEGATE_KEY),
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(
+                                                            accountAmount(sender, -amountToBeSent, false),
+                                                            accountAmount(receiver, amountToBeSent, false))
+                                                    .build(),
+                                            EMPTY_TUPLE_ARRAY)
+                                    .payingWith(GENESIS)
+                                    .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
+                                    .via(failedTransferFromTxn)
+                                    .gas(GAS_TO_OFFER));
+                }),
+                getTxnRecord(failedTransferFromTxn).andAllChildRecords().logged(),
+                childRecordsCheck(
                         failedTransferFromTxn,
                         CONTRACT_REVERT_EXECUTED,
                         recordWith()
@@ -1401,47 +1274,37 @@ public class AtomicCryptoTransferHTSSuite {
     @HapiTest
     final Stream<DynamicTest> delegateCallShouldFail() {
         final var failedTransferFromTxn = "failed_txn";
-        final long allowance = 20L;
 
-        return propertyPreservingHapiSpec(
-                        "delegateCallShouldFail",
-                        NONDETERMINISTIC_FUNCTION_PARAMETERS,
-                        ACCEPTED_MONO_GAS_CALCULATION_DIFFERENCE,
-                        NONDETERMINISTIC_TRANSACTION_FEES)
-                .preserving("contracts.precompile.atomicCryptoTransfer.enabled")
-                .given(
-                        overriding("contracts.precompile.atomicCryptoTransfer.enabled", "true"),
-                        cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS),
-                        uploadInitCode(CONTRACT),
-                        contractCreate(CONTRACT))
-                .when(
-                        withOpContext((spec, opLog) -> {
-                            final var sender = spec.registry().getAccountID(SENDER);
-                            final var receiver = spec.registry().getAccountID(RECEIVER);
-                            final var amountToBeSent = 50 * ONE_HBAR;
+        return hapiTest(
+                cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
+                cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS),
+                uploadInitCode(CONTRACT),
+                contractCreate(CONTRACT),
+                withOpContext((spec, opLog) -> {
+                    final var sender = spec.registry().getAccountID(SENDER);
+                    final var receiver = spec.registry().getAccountID(RECEIVER);
+                    final var amountToBeSent = 50 * ONE_HBAR;
 
-                            allRunFor(
-                                    spec,
-                                    newKeyNamed(DELEGATE_KEY)
-                                            .shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
-                                    cryptoUpdate(SENDER).key(DELEGATE_KEY),
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS_DELEGATE_CALL,
-                                                    transferList()
-                                                            .withAccountAmounts(
-                                                                    accountAmount(sender, -amountToBeSent, false),
-                                                                    accountAmount(receiver, amountToBeSent, false))
-                                                            .build(),
-                                                    EMPTY_TUPLE_ARRAY)
-                                            .payingWith(GENESIS)
-                                            .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
-                                            .via(failedTransferFromTxn)
-                                            .gas(GAS_TO_OFFER));
-                        }),
-                        getTxnRecord(failedTransferFromTxn).andAllChildRecords().logged())
-                .then(childRecordsCheck(failedTransferFromTxn, CONTRACT_REVERT_EXECUTED));
+                    allRunFor(
+                            spec,
+                            newKeyNamed(DELEGATE_KEY).shape(DELEGATE_CONTRACT_KEY_SHAPE.signedWith(sigs(ON, CONTRACT))),
+                            cryptoUpdate(SENDER).key(DELEGATE_KEY),
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS_DELEGATE_CALL,
+                                            transferList()
+                                                    .withAccountAmounts(
+                                                            accountAmount(sender, -amountToBeSent, false),
+                                                            accountAmount(receiver, amountToBeSent, false))
+                                                    .build(),
+                                            EMPTY_TUPLE_ARRAY)
+                                    .payingWith(GENESIS)
+                                    .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
+                                    .via(failedTransferFromTxn)
+                                    .gas(GAS_TO_OFFER));
+                }),
+                getTxnRecord(failedTransferFromTxn).andAllChildRecords().logged(),
+                childRecordsCheck(failedTransferFromTxn, CONTRACT_REVERT_EXECUTED));
     }
 
     @HapiTest
@@ -1507,69 +1370,59 @@ public class AtomicCryptoTransferHTSSuite {
                                                 .withStatus(INVALID_FULL_PREFIX_SIGNATURE_FOR_PRECOMPILE)))));
     }
 
-    @HapiTest
+    @LeakyHapiTest(overrides = {"contracts.permittedDelegateCallers"})
     final Stream<DynamicTest> blockCryptoTransferForPermittedDelegates() {
         final var blockCryptoTransferForPermittedDelegates = "blockCryptoTransferForPermittedDelegates";
         final AtomicLong whitelistedCalleeMirrorNum = new AtomicLong();
         final AtomicReference<String> whitelistedCalleeMirrorAddr = new AtomicReference<>();
 
-        return propertyPreservingHapiSpec(
-                        "blockCryptoTransferForPermittedDelegates",
-                        NONDETERMINISTIC_FUNCTION_PARAMETERS,
-                        NONDETERMINISTIC_TRANSACTION_FEES,
-                        NONDETERMINISTIC_NONCE)
-                .preserving("contracts.precompile.atomicCryptoTransfer.enabled", CONTRACTS_PERMITTED_DELEGATE_CALLERS)
-                .given(
-                        cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(TOKEN_TREASURY),
-                        tokenCreate(FUNGIBLE_TOKEN)
-                                .tokenType(TokenType.FUNGIBLE_COMMON)
-                                .initialSupply(TOTAL_SUPPLY)
-                                .treasury(TOKEN_TREASURY),
-                        tokenAssociate(SENDER, List.of(FUNGIBLE_TOKEN)),
-                        tokenAssociate(RECEIVER, List.of(FUNGIBLE_TOKEN)),
-                        cryptoTransfer(moving(200, FUNGIBLE_TOKEN).between(TOKEN_TREASURY, SENDER)),
-                        uploadInitCode(CONTRACT),
-                        contractCreate(CONTRACT).adminKey(DEFAULT_PAYER).exposingNumTo(num -> {
-                            whitelistedCalleeMirrorNum.set(num);
-                            whitelistedCalleeMirrorAddr.set(asHexedSolidityAddress(0, 0, num));
-                        }))
-                .when(
-                        overridingTwo(
-                                "contracts.precompile.atomicCryptoTransfer.enabled",
-                                "true",
-                                CONTRACTS_PERMITTED_DELEGATE_CALLERS,
-                                String.valueOf(whitelistedCalleeMirrorNum.get())),
-                        withOpContext((spec, opLog) -> {
-                            final var token = spec.registry().getTokenID(FUNGIBLE_TOKEN);
-                            final var sender = spec.registry().getAccountID(SENDER);
-                            final var receiver = spec.registry().getAccountID(RECEIVER);
-                            final var amountToBeSent = 50L;
+        return hapiTest(
+                cryptoCreate(SENDER).balance(10 * ONE_HUNDRED_HBARS),
+                cryptoCreate(RECEIVER).balance(2 * ONE_HUNDRED_HBARS),
+                cryptoCreate(TOKEN_TREASURY),
+                tokenCreate(FUNGIBLE_TOKEN)
+                        .tokenType(TokenType.FUNGIBLE_COMMON)
+                        .initialSupply(TOTAL_SUPPLY)
+                        .treasury(TOKEN_TREASURY),
+                tokenAssociate(SENDER, List.of(FUNGIBLE_TOKEN)),
+                tokenAssociate(RECEIVER, List.of(FUNGIBLE_TOKEN)),
+                cryptoTransfer(moving(200, FUNGIBLE_TOKEN).between(TOKEN_TREASURY, SENDER)),
+                uploadInitCode(CONTRACT),
+                contractCreate(CONTRACT).adminKey(DEFAULT_PAYER).exposingNumTo(num -> {
+                    whitelistedCalleeMirrorNum.set(num);
+                    whitelistedCalleeMirrorAddr.set(asHexedSolidityAddress(0, 0, num));
+                }),
+                sourcing(() -> overriding(
+                        "contracts.permittedDelegateCallers", String.valueOf(whitelistedCalleeMirrorNum.get()))),
+                withOpContext((spec, opLog) -> {
+                    final var token = spec.registry().getTokenID(FUNGIBLE_TOKEN);
+                    final var sender = spec.registry().getAccountID(SENDER);
+                    final var receiver = spec.registry().getAccountID(RECEIVER);
+                    final var amountToBeSent = 50L;
 
-                            allRunFor(
-                                    spec,
-                                    contractCall(
-                                                    CONTRACT,
-                                                    TRANSFER_MULTIPLE_TOKENS,
-                                                    transferList()
-                                                            .withAccountAmounts(EMPTY_TUPLE_ARRAY)
-                                                            .build(),
-                                                    wrapIntoTupleArray(tokenTransferList()
-                                                            .forToken(token)
-                                                            .withAccountAmounts(
-                                                                    accountAmount(sender, -amountToBeSent, false),
-                                                                    accountAmount(receiver, amountToBeSent, false))
-                                                            .build()))
-                                            .payingWith(DEFAULT_PAYER)
-                                            .via(blockCryptoTransferForPermittedDelegates)
-                                            .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
-                                            .gas(GAS_TO_OFFER));
-                        }),
-                        getTxnRecord(blockCryptoTransferForPermittedDelegates)
-                                .andAllChildRecords()
-                                .logged())
-                .then(childRecordsCheck(
+                    allRunFor(
+                            spec,
+                            contractCall(
+                                            CONTRACT,
+                                            TRANSFER_MULTIPLE_TOKENS,
+                                            transferList()
+                                                    .withAccountAmounts(EMPTY_TUPLE_ARRAY)
+                                                    .build(),
+                                            wrapIntoTupleArray(tokenTransferList()
+                                                    .forToken(token)
+                                                    .withAccountAmounts(
+                                                            accountAmount(sender, -amountToBeSent, false),
+                                                            accountAmount(receiver, amountToBeSent, false))
+                                                    .build()))
+                                    .payingWith(DEFAULT_PAYER)
+                                    .via(blockCryptoTransferForPermittedDelegates)
+                                    .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
+                                    .gas(GAS_TO_OFFER));
+                }),
+                getTxnRecord(blockCryptoTransferForPermittedDelegates)
+                        .andAllChildRecords()
+                        .logged(),
+                childRecordsCheck(
                         blockCryptoTransferForPermittedDelegates,
                         CONTRACT_REVERT_EXECUTED,
                         recordWith()
