+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/hip906/HbarAllowanceApprovalTest.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.hedera.services.bdd.suites.contract.precompile;
+package com.hedera.services.bdd.suites.hip906;
 
 import static com.hedera.services.bdd.junit.TestTags.SMART_CONTRACT;
 import static com.hedera.services.bdd.spec.HapiPropertySource.idAsHeadlongAddress;
-import static com.hedera.services.bdd.spec.HapiSpec.propertyPreservingHapiSpec;
+import static com.hedera.services.bdd.spec.HapiSpec.hapiTest;
 import static com.hedera.services.bdd.spec.assertions.ContractFnResultAsserts.isLiteralResult;
 import static com.hedera.services.bdd.spec.assertions.ContractFnResultAsserts.resultWith;
 import static com.hedera.services.bdd.spec.assertions.TransactionRecordAsserts.recordWith;
@@ -34,7 +34,6 @@ import static com.hedera.services.bdd.spec.transactions.TxnVerbs.uploadInitCode;
 import static com.hedera.services.bdd.spec.transactions.contract.HapiParserUtil.asHeadlongAddress;
 import static com.hedera.services.bdd.spec.transactions.crypto.HapiCryptoTransfer.tinyBarsFromTo;
 import static com.hedera.services.bdd.spec.utilops.CustomSpecAssert.allRunFor;
-import static com.hedera.services.bdd.spec.utilops.UtilVerbs.overriding;
 import static com.hedera.services.bdd.spec.utilops.UtilVerbs.withOpContext;
 import static com.hedera.services.bdd.suites.HapiSuite.ONE_HUNDRED_HBARS;
 import static com.hedera.services.bdd.suites.contract.Utils.getABIFor;
@@ -46,15 +45,17 @@ import com.hedera.services.bdd.junit.HapiTest;
 import java.math.BigInteger;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Stream;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.Tag;
 
+/**
+ * Tests expected behavior of the HRC-632 {@code hbarApprove(address spender, int256 amount)} and
+ * {@code hbarAllowance(address spender)} functions when the {@code contracts.systemContract.accountService.enabled}
+ * feature flag is on for <a href="https://hips.hedera.com/hip/hip-906">HIP-906</a> (which is true by default in
+ * the current release.)
+ */
 @Tag(SMART_CONTRACT)
-public class HbarAllowanceApprovalSuite {
-    private static final Logger log = LogManager.getLogger(HbarAllowanceApprovalSuite.class);
-    private static final String MULTI_KEY = "multikey";
+public class HbarAllowanceApprovalTest {
     private static final String SPENDER = "spender";
     private static final String HRC632_CONTRACT = "HRC632Contract";
     private static final String IHRC632 = "IHRC632";
@@ -67,24 +68,19 @@ public class HbarAllowanceApprovalSuite {
     private static final String HBAR_ALLOWANCE_CALL = "hbarAllowanceCall";
     private static final String HBAR_APPROVE_CALL = "hbarApproveCall";
     private static final String HBAR_APPROVE_DELEGATE_CALL = "hbarApproveDelegateCall";
-    private static final String CONTRACTS_SYSTEM_CONTRACT_ACCOUNT_SERVICE_ENABLED =
-            "contracts.systemContract.accountService.enabled";
 
     @HapiTest
     final Stream<DynamicTest> hrc632AllowanceFromEOA() {
         final AtomicReference<Address> accountNum = new AtomicReference<>();
         final AtomicReference<Address> spenderNum = new AtomicReference<>();
 
-        return propertyPreservingHapiSpec("hrc632AllowanceFromEOA")
-                .preserving(CONTRACTS_SYSTEM_CONTRACT_ACCOUNT_SERVICE_ENABLED)
-                .given(
-                        overriding(CONTRACTS_SYSTEM_CONTRACT_ACCOUNT_SERVICE_ENABLED, "true"),
-                        cryptoCreate(ACCOUNT)
-                                .balance(100 * ONE_HUNDRED_HBARS)
-                                .exposingCreatedIdTo(id -> accountNum.set(idAsHeadlongAddress(id))),
-                        cryptoCreate(SPENDER).exposingCreatedIdTo(id -> spenderNum.set(idAsHeadlongAddress(id))),
-                        cryptoApproveAllowance().payingWith(ACCOUNT).addCryptoAllowance(ACCOUNT, SPENDER, 1_000_000))
-                .when(withOpContext((spec, opLog) -> {
+        return hapiTest(
+                cryptoCreate(ACCOUNT)
+                        .balance(100 * ONE_HUNDRED_HBARS)
+                        .exposingCreatedIdTo(id -> accountNum.set(idAsHeadlongAddress(id))),
+                cryptoCreate(SPENDER).exposingCreatedIdTo(id -> spenderNum.set(idAsHeadlongAddress(id))),
+                cryptoApproveAllowance().payingWith(ACCOUNT).addCryptoAllowance(ACCOUNT, SPENDER, 1_000_000),
+                withOpContext((spec, opLog) -> {
                     var accountAddress = "0.0." + accountNum.get().value();
                     var spenderAddress = spenderNum.get();
                     allRunFor(
@@ -100,8 +96,8 @@ public class HbarAllowanceApprovalSuite {
                                     .payingWith(ACCOUNT)
                                     .gas(1_000_000)
                                     .via(HBAR_ALLOWANCE_TXN));
-                }))
-                .then(getTxnRecord(HBAR_ALLOWANCE_TXN)
+                }),
+                getTxnRecord(HBAR_ALLOWANCE_TXN)
                         .logged()
                         .hasPriority(recordWith()
                                 .status(SUCCESS)
@@ -122,15 +118,12 @@ public class HbarAllowanceApprovalSuite {
         final AtomicReference<Address> accountNum = new AtomicReference<>();
         final AtomicReference<Address> spenderNum = new AtomicReference<>();
 
-        return propertyPreservingHapiSpec("hrc632ApproveFromEOA")
-                .preserving(CONTRACTS_SYSTEM_CONTRACT_ACCOUNT_SERVICE_ENABLED)
-                .given(
-                        overriding(CONTRACTS_SYSTEM_CONTRACT_ACCOUNT_SERVICE_ENABLED, "true"),
-                        cryptoCreate(ACCOUNT)
-                                .balance(100 * ONE_HUNDRED_HBARS)
-                                .exposingCreatedIdTo(id -> accountNum.set(idAsHeadlongAddress(id))),
-                        cryptoCreate(SPENDER).exposingCreatedIdTo(id -> spenderNum.set(idAsHeadlongAddress(id))))
-                .when(withOpContext((spec, opLog) -> {
+        return hapiTest(
+                cryptoCreate(ACCOUNT)
+                        .balance(100 * ONE_HUNDRED_HBARS)
+                        .exposingCreatedIdTo(id -> accountNum.set(idAsHeadlongAddress(id))),
+                cryptoCreate(SPENDER).exposingCreatedIdTo(id -> spenderNum.set(idAsHeadlongAddress(id))),
+                withOpContext((spec, opLog) -> {
                     var accountAddress = "0.0." + accountNum.get().value();
                     var spenderAddress = spenderNum.get();
                     allRunFor(
@@ -147,8 +140,8 @@ public class HbarAllowanceApprovalSuite {
                                     .payingWith(ACCOUNT)
                                     .gas(1_000_000)
                                     .via(HBAR_APPROVE_TXN));
-                }))
-                .then(getTxnRecord(HBAR_APPROVE_TXN)
+                }),
+                getTxnRecord(HBAR_APPROVE_TXN)
                         .logged()
                         .hasPriority(recordWith()
                                 .status(SUCCESS)
@@ -167,20 +160,17 @@ public class HbarAllowanceApprovalSuite {
         final AtomicReference<Address> accountNum = new AtomicReference<>();
         final AtomicReference<Address> spenderNum = new AtomicReference<>();
 
-        return propertyPreservingHapiSpec("hrc632AllowanceFromContract")
-                .preserving(CONTRACTS_SYSTEM_CONTRACT_ACCOUNT_SERVICE_ENABLED)
-                .given(
-                        overriding(CONTRACTS_SYSTEM_CONTRACT_ACCOUNT_SERVICE_ENABLED, "true"),
-                        cryptoCreate(ACCOUNT)
-                                .balance(100 * ONE_HUNDRED_HBARS)
-                                .exposingCreatedIdTo(id -> accountNum.set(idAsHeadlongAddress(id))),
-                        cryptoCreate(SPENDER).exposingCreatedIdTo(id -> spenderNum.set(idAsHeadlongAddress(id))),
-                        uploadInitCode(HRC632_CONTRACT),
-                        contractCreate(HRC632_CONTRACT),
-                        cryptoApproveAllowance()
-                                .addCryptoAllowance(ACCOUNT, SPENDER, 1_000_000)
-                                .payingWith(ACCOUNT))
-                .when(withOpContext((spec, opLog) -> {
+        return hapiTest(
+                cryptoCreate(ACCOUNT)
+                        .balance(100 * ONE_HUNDRED_HBARS)
+                        .exposingCreatedIdTo(id -> accountNum.set(idAsHeadlongAddress(id))),
+                cryptoCreate(SPENDER).exposingCreatedIdTo(id -> spenderNum.set(idAsHeadlongAddress(id))),
+                uploadInitCode(HRC632_CONTRACT),
+                contractCreate(HRC632_CONTRACT),
+                cryptoApproveAllowance()
+                        .addCryptoAllowance(ACCOUNT, SPENDER, 1_000_000)
+                        .payingWith(ACCOUNT),
+                withOpContext((spec, opLog) -> {
                     var spenderAddress = spenderNum.get();
                     var ownerAddress = accountNum.get();
                     allRunFor(
@@ -190,8 +180,8 @@ public class HbarAllowanceApprovalSuite {
                                     .payingWith(ACCOUNT)
                                     .gas(1_000_000)
                                     .via(HBAR_ALLOWANCE_TXN));
-                }))
-                .then(getTxnRecord(HBAR_ALLOWANCE_TXN)
+                }),
+                getTxnRecord(HBAR_ALLOWANCE_TXN)
                         .logged()
                         .hasPriority(recordWith()
                                 .status(SUCCESS)
@@ -213,20 +203,16 @@ public class HbarAllowanceApprovalSuite {
         final AtomicReference<Address> spenderNum = new AtomicReference<>();
         final AtomicReference<Address> contractNum = new AtomicReference<>();
 
-        return propertyPreservingHapiSpec("hrc632FromContract")
-                .preserving(CONTRACTS_SYSTEM_CONTRACT_ACCOUNT_SERVICE_ENABLED)
-                .given(
-                        overriding(CONTRACTS_SYSTEM_CONTRACT_ACCOUNT_SERVICE_ENABLED, "true"),
-                        cryptoCreate(ACCOUNT)
-                                .balance(100 * ONE_HUNDRED_HBARS)
-                                .exposingCreatedIdTo(id -> accountNum.set(idAsHeadlongAddress(id))),
-                        cryptoCreate(SPENDER).exposingCreatedIdTo(id -> spenderNum.set(idAsHeadlongAddress(id))),
-                        uploadInitCode(HRC632_CONTRACT),
-                        contractCreate(HRC632_CONTRACT),
-                        getContractInfo(HRC632_CONTRACT)
-                                .exposingEvmAddress(cb -> contractNum.set(asHeadlongAddress(cb))),
-                        cryptoTransfer(tinyBarsFromTo(ACCOUNT, HRC632_CONTRACT, 1_000_000L)))
-                .when(withOpContext((spec, opLog) -> {
+        return hapiTest(
+                cryptoCreate(ACCOUNT)
+                        .balance(100 * ONE_HUNDRED_HBARS)
+                        .exposingCreatedIdTo(id -> accountNum.set(idAsHeadlongAddress(id))),
+                cryptoCreate(SPENDER).exposingCreatedIdTo(id -> spenderNum.set(idAsHeadlongAddress(id))),
+                uploadInitCode(HRC632_CONTRACT),
+                contractCreate(HRC632_CONTRACT),
+                getContractInfo(HRC632_CONTRACT).exposingEvmAddress(cb -> contractNum.set(asHeadlongAddress(cb))),
+                cryptoTransfer(tinyBarsFromTo(ACCOUNT, HRC632_CONTRACT, 1_000_000L)),
+                withOpContext((spec, opLog) -> {
                     var contractAddress = contractNum.get();
                     var spenderAddress = spenderNum.get();
                     allRunFor(
@@ -246,33 +232,32 @@ public class HbarAllowanceApprovalSuite {
                                     .payingWith(ACCOUNT)
                                     .gas(1_000_000)
                                     .via(HBAR_ALLOWANCE_TXN));
-                }))
-                .then(
-                        getTxnRecord(HBAR_APPROVE_TXN)
-                                .logged()
-                                .hasPriority(recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .resultThruAbi(
-                                                        getABIFor(
-                                                                com.hedera.services.bdd.suites.contract.Utils
-                                                                        .FunctionType.FUNCTION,
-                                                                HBAR_APPROVE_CALL,
-                                                                HRC632_CONTRACT),
-                                                        isLiteralResult(new Object[] {Long.valueOf(22)})))),
-                        getTxnRecord(HBAR_ALLOWANCE_TXN)
-                                .logged()
-                                .hasPriority(recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .resultThruAbi(
-                                                        getABIFor(
-                                                                com.hedera.services.bdd.suites.contract.Utils
-                                                                        .FunctionType.FUNCTION,
-                                                                HBAR_ALLOWANCE_CALL,
-                                                                HRC632_CONTRACT),
-                                                        isLiteralResult(new Object[] {
-                                                            Long.valueOf(22), BigInteger.valueOf(1_000_000L)
+                }),
+                getTxnRecord(HBAR_APPROVE_TXN)
+                        .logged()
+                        .hasPriority(recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .resultThruAbi(
+                                                getABIFor(
+                                                        com.hedera.services.bdd.suites.contract.Utils.FunctionType
+                                                                .FUNCTION,
+                                                        HBAR_APPROVE_CALL,
+                                                        HRC632_CONTRACT),
+                                                isLiteralResult(new Object[] {Long.valueOf(22)})))),
+                getTxnRecord(HBAR_ALLOWANCE_TXN)
+                        .logged()
+                        .hasPriority(recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .resultThruAbi(
+                                                getABIFor(
+                                                        com.hedera.services.bdd.suites.contract.Utils.FunctionType
+                                                                .FUNCTION,
+                                                        HBAR_ALLOWANCE_CALL,
+                                                        HRC632_CONTRACT),
+                                                isLiteralResult(
+                                                        new Object[] {Long.valueOf(22), BigInteger.valueOf(1_000_000L)
                                                         })))));
     }
 
@@ -281,16 +266,13 @@ public class HbarAllowanceApprovalSuite {
         final AtomicReference<Address> accountNum = new AtomicReference<>();
         final AtomicReference<Address> spenderNum = new AtomicReference<>();
 
-        return propertyPreservingHapiSpec("hrc632ApproveFromEOAFailsWhenWrongKey")
-                .preserving(CONTRACTS_SYSTEM_CONTRACT_ACCOUNT_SERVICE_ENABLED)
-                .given(
-                        overriding(CONTRACTS_SYSTEM_CONTRACT_ACCOUNT_SERVICE_ENABLED, "true"),
-                        cryptoCreate(SIGNER).balance(100 * ONE_HUNDRED_HBARS),
-                        cryptoCreate(ACCOUNT)
-                                .balance(100 * ONE_HUNDRED_HBARS)
-                                .exposingCreatedIdTo(id -> accountNum.set(idAsHeadlongAddress(id))),
-                        cryptoCreate(SPENDER).exposingCreatedIdTo(id -> spenderNum.set(idAsHeadlongAddress(id))))
-                .when(withOpContext((spec, opLog) -> {
+        return hapiTest(
+                cryptoCreate(SIGNER).balance(100 * ONE_HUNDRED_HBARS),
+                cryptoCreate(ACCOUNT)
+                        .balance(100 * ONE_HUNDRED_HBARS)
+                        .exposingCreatedIdTo(id -> accountNum.set(idAsHeadlongAddress(id))),
+                cryptoCreate(SPENDER).exposingCreatedIdTo(id -> spenderNum.set(idAsHeadlongAddress(id))),
+                withOpContext((spec, opLog) -> {
                     var accountAddress = "0.0." + accountNum.get().value();
                     var spenderAddress = spenderNum.get();
                     allRunFor(
@@ -308,10 +290,8 @@ public class HbarAllowanceApprovalSuite {
                                     .gas(1_000_000)
                                     .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
                                     .via(HBAR_APPROVE_TXN));
-                }))
-                .then(getTxnRecord(HBAR_APPROVE_TXN)
-                        .logged()
-                        .hasPriority(recordWith().status(CONTRACT_REVERT_EXECUTED)));
+                }),
+                getTxnRecord(HBAR_APPROVE_TXN).logged().hasPriority(recordWith().status(CONTRACT_REVERT_EXECUTED)));
     }
 
     @HapiTest
@@ -320,20 +300,16 @@ public class HbarAllowanceApprovalSuite {
         final AtomicReference<Address> spenderNum = new AtomicReference<>();
         final AtomicReference<Address> contractNum = new AtomicReference<>();
 
-        return propertyPreservingHapiSpec("hrc632FromContract")
-                .preserving(CONTRACTS_SYSTEM_CONTRACT_ACCOUNT_SERVICE_ENABLED)
-                .given(
-                        overriding(CONTRACTS_SYSTEM_CONTRACT_ACCOUNT_SERVICE_ENABLED, "true"),
-                        cryptoCreate(ACCOUNT)
-                                .balance(100 * ONE_HUNDRED_HBARS)
-                                .exposingCreatedIdTo(id -> accountNum.set(idAsHeadlongAddress(id))),
-                        cryptoCreate(SPENDER).exposingCreatedIdTo(id -> spenderNum.set(idAsHeadlongAddress(id))),
-                        uploadInitCode(HRC632_CONTRACT),
-                        contractCreate(HRC632_CONTRACT),
-                        getContractInfo(HRC632_CONTRACT)
-                                .exposingEvmAddress(cb -> contractNum.set(asHeadlongAddress(cb))),
-                        cryptoTransfer(tinyBarsFromTo(ACCOUNT, HRC632_CONTRACT, 1_000_000L)))
-                .when(withOpContext((spec, opLog) -> {
+        return hapiTest(
+                cryptoCreate(ACCOUNT)
+                        .balance(100 * ONE_HUNDRED_HBARS)
+                        .exposingCreatedIdTo(id -> accountNum.set(idAsHeadlongAddress(id))),
+                cryptoCreate(SPENDER).exposingCreatedIdTo(id -> spenderNum.set(idAsHeadlongAddress(id))),
+                uploadInitCode(HRC632_CONTRACT),
+                contractCreate(HRC632_CONTRACT),
+                getContractInfo(HRC632_CONTRACT).exposingEvmAddress(cb -> contractNum.set(asHeadlongAddress(cb))),
+                cryptoTransfer(tinyBarsFromTo(ACCOUNT, HRC632_CONTRACT, 1_000_000L)),
+                withOpContext((spec, opLog) -> {
                     var accountAddress = accountNum.get();
                     var spenderAddress = spenderNum.get();
                     allRunFor(
@@ -354,25 +330,21 @@ public class HbarAllowanceApprovalSuite {
                                     .payingWith(ACCOUNT)
                                     .gas(1_000_000)
                                     .via(HBAR_ALLOWANCE_TXN));
-                }))
-                .then(
-                        getTxnRecord(HBAR_APPROVE_TXN)
-                                .logged()
-                                .hasPriority(recordWith().status(CONTRACT_REVERT_EXECUTED)),
-                        getTxnRecord(HBAR_ALLOWANCE_TXN)
-                                .logged()
-                                .hasPriority(recordWith()
-                                        .status(SUCCESS)
-                                        .contractCallResult(resultWith()
-                                                .resultThruAbi(
-                                                        getABIFor(
-                                                                com.hedera.services.bdd.suites.contract.Utils
-                                                                        .FunctionType.FUNCTION,
-                                                                HBAR_ALLOWANCE_CALL,
-                                                                HRC632_CONTRACT),
-                                                        isLiteralResult(
-                                                                new Object[] {Long.valueOf(22), BigInteger.valueOf(0)
-                                                                })))));
+                }),
+                getTxnRecord(HBAR_APPROVE_TXN).logged().hasPriority(recordWith().status(CONTRACT_REVERT_EXECUTED)),
+                getTxnRecord(HBAR_ALLOWANCE_TXN)
+                        .logged()
+                        .hasPriority(recordWith()
+                                .status(SUCCESS)
+                                .contractCallResult(resultWith()
+                                        .resultThruAbi(
+                                                getABIFor(
+                                                        com.hedera.services.bdd.suites.contract.Utils.FunctionType
+                                                                .FUNCTION,
+                                                        HBAR_ALLOWANCE_CALL,
+                                                        HRC632_CONTRACT),
+                                                isLiteralResult(
+                                                        new Object[] {Long.valueOf(22), BigInteger.valueOf(0)})))));
     }
 
     @HapiTest
@@ -381,20 +353,16 @@ public class HbarAllowanceApprovalSuite {
         final AtomicReference<Address> spenderNum = new AtomicReference<>();
         final AtomicReference<Address> contractNum = new AtomicReference<>();
 
-        return propertyPreservingHapiSpec("hrc632FromContract")
-                .preserving(CONTRACTS_SYSTEM_CONTRACT_ACCOUNT_SERVICE_ENABLED)
-                .given(
-                        overriding(CONTRACTS_SYSTEM_CONTRACT_ACCOUNT_SERVICE_ENABLED, "true"),
-                        cryptoCreate(ACCOUNT)
-                                .balance(100 * ONE_HUNDRED_HBARS)
-                                .exposingCreatedIdTo(id -> accountNum.set(idAsHeadlongAddress(id))),
-                        cryptoCreate(SPENDER).exposingCreatedIdTo(id -> spenderNum.set(idAsHeadlongAddress(id))),
-                        uploadInitCode(HRC632_CONTRACT),
-                        contractCreate(HRC632_CONTRACT),
-                        getContractInfo(HRC632_CONTRACT)
-                                .exposingEvmAddress(cb -> contractNum.set(asHeadlongAddress(cb))),
-                        cryptoTransfer(tinyBarsFromTo(ACCOUNT, HRC632_CONTRACT, 1_000_000L)))
-                .when(withOpContext((spec, opLog) -> {
+        return hapiTest(
+                cryptoCreate(ACCOUNT)
+                        .balance(100 * ONE_HUNDRED_HBARS)
+                        .exposingCreatedIdTo(id -> accountNum.set(idAsHeadlongAddress(id))),
+                cryptoCreate(SPENDER).exposingCreatedIdTo(id -> spenderNum.set(idAsHeadlongAddress(id))),
+                uploadInitCode(HRC632_CONTRACT),
+                contractCreate(HRC632_CONTRACT),
+                getContractInfo(HRC632_CONTRACT).exposingEvmAddress(cb -> contractNum.set(asHeadlongAddress(cb))),
+                cryptoTransfer(tinyBarsFromTo(ACCOUNT, HRC632_CONTRACT, 1_000_000L)),
+                withOpContext((spec, opLog) -> {
                     var contractAddress = contractNum.get();
                     var spenderAddress = spenderNum.get();
                     allRunFor(
@@ -410,9 +378,7 @@ public class HbarAllowanceApprovalSuite {
                                     .gas(1_000_000)
                                     .hasKnownStatus(CONTRACT_REVERT_EXECUTED)
                                     .via(HBAR_APPROVE_TXN));
-                }))
-                .then(getTxnRecord(HBAR_APPROVE_TXN)
-                        .logged()
-                        .hasPriority(recordWith().status(CONTRACT_REVERT_EXECUTED)));
+                }),
+                getTxnRecord(HBAR_APPROVE_TXN).logged().hasPriority(recordWith().status(CONTRACT_REVERT_EXECUTED)));
     }
 }
